### 2025-01-20 #7700 @editorialbot
[PRE REVIEW]: eo-tides: Tide modelling tools for large-scale satellite Earth observation analysis
**Submitting author:** @robbibt (<a href="http://orcid.org/0000-0002-1533-2599">Robbi Bishop-Taylor</a>)
**Repository:** https://github.com/GeoscienceAustralia/eo-tides
**Branch with paper.md** (empty if default branch):
**Version:** 0.5.0
