+++ paper/paper.bib
@@ -57,7 +57,8 @@
   number={11},
   pages={3417--3426},
   year={2012},
-  publisher={Molecular Diversity Preservation International (MDPI)}
+  publisher={Molecular Diversity Preservation International (MDPI)},
+  doi = {10.3390/rs4113417}
 }
 
 @article{sagar2017item,
@@ -67,7 +68,8 @@
   volume={195},
   pages={153--169},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.rse.2017.04.009}
 }
 
 @inproceedings{carrere2022new,
@@ -86,7 +88,7 @@
   number = {3},
   pages = {293-314},
   keywords = {sea level records, sea level rise, storm surges, storm tides, tide gauge},
-  doi = {https://doi.org/10.1002/gdj3.174},
+  doi = {10.1002/gdj3.174},
   url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/gdj3.174},
   eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/gdj3.174},
   abstract = {Abstract This paper describes a major update to the quasi-global, higher-frequency sea-level dataset known as GESLA (Global Extreme Sea Level Analysis). Versions 1 (released 2009) and 2 (released 2016) of the dataset have been used in many published studies, across a wide range of oceanographic and coastal engineering-related investigations concerned with evaluating tides, storm surges, extreme sea levels, and other related processes. The third version of the dataset (released 2021), presented here, contains double the number of years of data, and nearly four times the number of records, compared to Version 2. The dataset consists of records obtained from multiple sources around the world. This paper describes the assembly of the dataset, its processing, and its format, and outlines potential future improvements.},
@@ -132,7 +134,8 @@
   volume={143},
   pages={204--215},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.rse.2013.12.019}
 }
 
 @article{bishop2021mapping,
@@ -214,7 +217,8 @@
   volume={37},
   pages={101425},
   year={2025},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.2139/ssrn.4858713}
 }
 
 @misc{stac2024,
