### 2025-01-20 #7700 @robbibt
Thanks @kthyng, I'll work on this now. Here are some possible suggestions for reviewers with expertise in Python packages designed for coastal analysis:

* Kilian Vos (kvos), author `CoastSat` Python package
* Sharon Fitzpatrick (2320sharon), author `CoastSeg` Python package
* Floris Calkoen (FlorisCalkoen), author `coastpy` Python package
* Freya Muir (fmemuir), author `vegsat` Python package
* Nicholas Pucino (npucino), author `sandpyper` Python package
