### 2025-01-21 7887c6d
First attempt at reducing word count
