+++ paper/paper.md
@@ -72,7 +72,7 @@ The [`eo_tides.model`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_ti
 
 `pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and earth tides. Tides are frequently decomposed into harmonic constants (or constituents) associated with the relative positions of the sun, moon and Earth. For ocean tides, `pyTMD.io` contains routines for reading major constituent values from commonly available tide models, and interpolating those values to spatial locations. Information for each of the supported tide models is stored within a JSON database that is supplied with `pyTMD`. `pyTMD.astro` contains routines for computing the positions of celestial bodies for a given time. Namely for ocean tides, `pyTMD` computes the longitudes of the sun (S), moon (H), lunar perigree (P), ascending lunar node (N) and solar perigree (PP). `pyTMD.arguments` contains routines for combining the astronomical coefficients with the "Doodson number" of each constituent, along with routines for adjusting the amplitude and phase of each constituent based on their modulations over the 18.6 year nodal period. Finally, `pyTMD.predict` uses results from those underlying functions to predict tidal values at a given location and time.
 
-To support integration with satellite EO data, the `model_tides` function from `eo_tides.model` wraps `pyTMD` functionality to return predicted tides in a standardised `pandas.DataFrame` format containing information about the tide model, location and time period of each modelled tide. This allows large analyses to be broken into smaller discrete chunks that can be processed in parallel before being combined as a final step. Parallelisation in `eo-tides` is automatically optimised based on the number of available workers and the number of requested tide models and tide modelling locations. This built-in parallelisation can significantly improve tide modelling performance, especially when run on a large multi-core machine (\autoref{tab:benchmark}).
+The `model_tides` function from `eo_tides.model` wraps `pyTMD` functionality to return tide predictions in a standardised `pandas.DataFrame` format, enabling integration with satellite EO data and parallelized processing for improved performance. Parallelisation in `eo-tides` is automatically optimised based on available workers and requested tide models and tide modelling locations. This built-in parallelisation can significantly improve tide modelling performance, especially for large-scale analyses run on a large multi-core machine (\autoref{tab:benchmark}). Additional functions like `model_phases` classify tides or determine flow/ebb tides, critical for interpreting satellite-observed coastal processes like changing turbidity and ocean colour [@sent2025time].
 
 Table: A [benchmark comparison](https://github.com/GeoscienceAustralia/eo-tides/blob/main/paper/benchmarking.ipynb) of tide modelling performance with parallelisation on vs. off. This comparison was performed on an 8-core and 32-core Linux machine, for a typical large-scale analysis involving a month of hourly tides modelled at 10,000 modelling locations using three tide models (FES2022, TPXO10, GOT5.6). \label{tab:benchmark}
 
@@ -81,15 +81,11 @@ Table: A [benchmark comparison](https://github.com/GeoscienceAustralia/eo-tides/
 | 8     | 2min 46s ± 663 ms | 9min 28s ± 536 ms  | 3.4x    |
 | 32    | 55.9 s ± 560 ms   | 9min 24s ± 749 ms  | 10.1x   |
 
[…]
