### 2025-01-21 80f0e20
Simplify
