+++ paper/paper.md
@@ -58,7 +58,7 @@ The `eo-tides` package aims to address these challenges by providing a set of pe
 
 ## Setting up tide models
 
-The [`eo_tides.utils`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.utils) module simplifies the setup of global ocean tide models, addressing a common barrier in coastal EO workflows. Tools like `list_models` provide feedback on available and supported models (\autoref{fig:list})., while `clip_models` can be used to improve performance by clipping large model files to smaller regions of interest, significantly reducing processing times for high-resolution models like FES2022.
+The [`eo_tides.utils`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.utils) module simplifies the setup of global ocean tide models, addressing a common barrier in coastal EO workflows. Tools like `list_models` provide feedback on available and supported models (\autoref{fig:list}), while `clip_models` can be used to improve performance by clipping large model files to smaller regions of interest, significantly reducing processing times for high-resolution models like FES2022.
 
 Comprehensive documentation is available to [guide users in setting up commonly used tide models](https://geoscienceaustralia.github.io/eo-tides/setup/), including downloading, uncompressing, and organizing data files.
 
@@ -70,9 +70,9 @@ The [`eo_tides.model`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_ti
 
 `pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and earth tides. Tides are frequently decomposed into harmonic constants (or constituents) associated with the relative positions of the sun, moon and Earth. For ocean tides, `pyTMD.io` contains routines for reading major constituent values from commonly available tide models, and interpolating those values to spatial locations. Information for each of the supported tide models is stored within a JSON database that is supplied with `pyTMD`. `pyTMD.astro` contains routines for computing the positions of celestial bodies for a given time. Namely for ocean tides, `pyTMD` computes the longitudes of the sun (S), moon (H), lunar perigree (P), ascending lunar node (N) and solar perigree (PP). `pyTMD.arguments` contains routines for combining the astronomical coefficients with the "Doodson number" of each constituent, along with routines for adjusting the amplitude and phase of each constituent based on their modulations over the 18.6 year nodal period. Finally, `pyTMD.predict` uses results from those underlying functions to predict tidal values at a given location and time.
 
-The `model_tides` function from `eo_tides.model` wraps `pyTMD` functionality to return tide predictions in a standardised `pandas.DataFrame` format, enabling integration with satellite EO data and parallelized processing for improved performance. Parallelisation in `eo-tides` is automatically optimised based on available workers and requested tide models and tide modelling locations. This built-in parallelisation can significantly improve tide modelling performance, especially for large-scale analyses run on a large multi-core machine (\autoref{tab:benchmark}). Additional functions like `model_phases` classify tides or determine flow/ebb tides, critical for interpreting satellite-observed coastal processes like changing turbidity and ocean colour [@sent2025time].
[…]
