### 2025-01-21 2d62cf3
Simplify
