### 2025-01-21 9b61fcd
Simplification
