+++ paper/paper.md
@@ -50,9 +50,7 @@ Satellite remote sensing offers an unparalleled method to examine dynamic coasta
 
 Conversely, information about ocean tides can also provide unique environmental insights that can greatly enhance the utility of coastal EO data. Conventionally, satellite data dimensions consider the geographical "where" and the temporal "when" of data acquisition. The addition of tide height as a new analysis dimension allows data to be filtered, sorted and analysed with respect to tidal processes, delivering a powerful re-imagining of traditional multi-temporal EO data analysis [@sagar2017item]. For example, satellite data can be analysed to focus on specific ecologically-significant tidal stages (e.g. high, low tide, spring or neap tides) or on particular tidal processes (e.g. ebb or flow tides; @sent2025time).
 
-This concept has been used to map tidally-corrected annual coastlines from Landsat satellite data at continental scale [@bishop2021mapping], generate maps of the extent and elevation of the intertidal zone [@murray2012continental; @sagar2017item; @bishop2019NIDEM], and create tidally-constrained imagery composites of the coastline [@sagar2018composites]. However, these approaches have been historically based on bespoke, closed-source or difficult to install tide modelling tools, limiting the reproducibility and portability of these techniques to new coastal EO applications. To support the next generation of coastal EO workflows, there is a pressing need for new open-source tools for combining satellite data with tide modelling.
-
-`eo-tides` aims to address these challenges by providing a set of performant open-source Python tools for attributing satellite EO data with modelled ocean tides. This functionality is provided in five main analysis modules (`utils`, `model`, `eo`, `stats`, `validation`) described briefly below.
+This concept has been used to map tidally-corrected annual coastlines from Landsat satellite data at continental scale [@bishop2021mapping], generate maps of the extent and elevation of the intertidal zone [@murray2012continental; @sagar2017item; @bishop2019NIDEM], and create tidally-constrained imagery composites of the coastline [@sagar2018composites]. However, these approaches have been historically based on bespoke, closed-source or difficult to install tide modelling tools, limiting the reproducibility and portability of these techniques to new coastal EO applications. To support the next generation of coastal EO workflows, there is a pressing need for new open-source tools for combining satellite data with tide modelling. `eo-tides` aims to address these challenges by providing a set of performant open-source Python tools for attributing satellite EO data with modelled ocean tides. This functionality is provided in five main analysis modules (`utils`, `model`, `eo`, `stats`, `validation`) described briefly below.
 
 # Features
 
@@ -98,20 +96,20 @@ Table: Comparison of the `tag_tides` and `pixel_tides` functions. \label{tab:tid
 
 The [`eo_tides.stats`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.stats) module identifies biases caused by complex tide alaising interactions interactions between tidal dynamics and satellite observations. These interactions can prevent satellites from observing the entire tide cycle [@eleveld2014estuarine; @sent2025time], and cause coastal EO studies to produce biased or misleading results [@bishop2019NIDEM]. The module produces a range of useful statistics that summarise how well a satellite time series captures real-world tidal conditions, include spread (coverage of tide range) and high/low-tide offsets (missed tidal extremes). Automated reports and plots provide insights further insights into potential biases affecting the analysis.
 
[…]
