### 2025-01-21 d973503
Minor wording fixes
