+++ paper/paper.md
@@ -64,7 +64,7 @@ The [`eo_tides.utils`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_ti
 
 The [`eo_tides.model`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.model) module is powered by advanced tide modelling functionality from the `pyTMD` Python package [@pytmd].
 
-`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and earth tides. Tides are frequently decomposed into harmonic constants (or constituents) associated with the relative positions of the sun, moon and Earth. `pyTMD.io` contains routines for reading major constituent values from commonly available ocean tide models, and interpolating those values to spatial locations. `pyTMD.astro` contains routines for computing the positions of celestial bodies for a given time. For ocean tides, `pyTMD` computes the longitudes of the sun (S), moon (H), lunar perigree (P), ascending lunar node (N) and solar perigree (PP). `pyTMD.arguments` contains routines for combining the astronomical coefficients with the "Doodson number" of each constituent, along with routines for adjusting the amplitude and phase of each constituent based on their modulations over the 18.6 year nodal period. Finally, `pyTMD.predict` uses results from those underlying functions to predict tidal values at a given location and time.
+`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and earth tides. Tides are frequently decomposed into harmonic constants (or constituents) associated with the relative positions of the sun, moon and Earth. `pyTMD.io` contains routines for reading major constituent values from commonly available ocean tide models, and interpolating those values spatially. `pyTMD.astro` contains routines for computing the positions of celestial bodies for a given time. For ocean tides, `pyTMD` computes the longitudes of the sun (S), moon (H), lunar perigree (P), ascending lunar node (N) and solar perigree (PP). `pyTMD.arguments` combines astronomical coefficients with the "Doodson number" of each constituent, and adjusts the amplitude and phase of each constituent based on their modulations over the 18.6 year nodal period. Finally, `pyTMD.predict` uses results from those underlying functions to predict tidal values at a given location and time.
 
 The `model_tides` function from `eo_tides.model` wraps `pyTMD` functionality to return tide predictions in a standardised `pandas.DataFrame` format, enabling integration with satellite EO data and parallelized processing for improved performance. Parallelisation in `eo-tides` is automatically optimised based on available workers and requested tide models and tide modelling locations. This parallelisation can significantly improve tide modelling performance, especially for large-scale analyses run on a multi-core machine (\autoref{tab:benchmark}). Additional functions like `model_phases` classify high, low or flow/ebb tides, critical for interpreting satellite-observed coastal processes like changing turbidity and ocean colour [@sent2025time].
 
@@ -77,18 +77,16 @@ Table: A [benchmark comparison](https://github.com/GeoscienceAustralia/eo-tides/
 
 ## Combining tides with satellite data
 
-The [`eo_tides.eo`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.eo) module integrates modelled tides with `xarray`-format satellite data. For tide attribution, `eo-tides` offers two approaches that differ in complexity and performance: `tag_tides` assigns a single tide height per timestep for small-scale studies, while `pixel_tides` models tides spatially and temporally for larger-scale analyses, producing a unique tide height for each pixel in a dataset (\autoref{tab:tide_stats}).
-
-These functions can be applied to free and open satellite data for any coastal or ocean location on the planet, for example using data loaded from the cloud using the [Open Data Cube](https://www.opendatacube.org/) and SpatioTemporal Asset Catalogue [@stac2024].
[…]
