+++ paper/paper.md
@@ -50,7 +50,7 @@ Satellite remote sensing offers an unparalleled method to examine dynamic coasta
 
 Conversely, information about ocean tides can also provide unique environmental insights that can greatly enhance the utility of coastal EO data. Conventionally, satellite data dimensions consider the geographical "where" and the temporal "when" of data acquisition. The addition of tide height as a new analysis dimension allows data to be filtered, sorted and analysed with respect to tidal processes, delivering a powerful re-imagining of traditional multi-temporal EO data analysis [@sagar2017item]. For example, satellite data can be analysed to focus on specific ecologically-significant tidal stages (e.g. high, low tide, spring or neap tides) or on particular tidal processes (e.g. ebb or flow tides; @sent2025time).
 
-This concept has been used to map tidally-corrected annual coastlines from Landsat satellite data at continental scale [@bishop2021mapping], generate maps of the extent and elevation of the intertidal zone [@murray2012continental; @sagar2017item; @bishop2019NIDEM], and create tidally-constrained imagery composites of the coastline [@sagar2018composites]. However, these approaches have been historically based on bespoke, closed-source or difficult to install tide modelling tools, limiting the reproducibility and portability of these techniques to new coastal EO applications. To support the next generation of coastal EO workflows, there is a pressing need for new open-source tools for combining satellite data with tide modelling. `eo-tides` aims to address these challenges through functionality provided in five main analysis modules (`utils`, `model`, `eo`, `stats`, `validation`) which are described briefly below.
+This concept has been used to map tidally-corrected annual coastlines from Landsat satellite data at continental scale [@bishop2021mapping], generate maps of the extent and elevation of the intertidal zone [@murray2012continental; @sagar2017item; @bishop2019NIDEM], and create tidally-constrained imagery composites of the coastline [@sagar2018composites]. However, these approaches have been historically based on bespoke, closed-source or difficult to install tide modelling tools, limiting the reproducibility and portability of these techniques to new coastal EO applications. To support the next generation of coastal EO workflows, there is a pressing need for new open-source tools for combining satellite data with tide modelling. `eo-tides` aims to address this need through functionality provided in five main analysis modules (`utils`, `model`, `eo`, `stats`, `validation`) which are described briefly below.
 
 # Features
 
@@ -84,7 +84,7 @@ Table: Comparison of the `tag_tides` and `pixel_tides` functions. \label{tab:tid
 | `tag_tides`                                                              | `pixel_tides`                                                                                   |
 | ------------------------------------------------------------------------ | ----------------------------------------------------------------------------------------------- |
 | - Assigns a single tide height to each satellite image timestep          | - Assigns a tide height to every individual pixel through time to capture spatial tide dynamics |
-| - Ideal for local or site-scale analysis                                 | - Ideal for regional to global-scale coastal product generation                                 |
+| - Ideal for local or site-scale analysis                                 | - Ideal for large-scale coastal product generation                                              |
 | - Fast, low memory use                                                   | - Slower, higher memory use                                                                     |
 | - Single tide height per image can produce artefacts and discontinuities | - Produce spatially seamless results across large extents                                       |
 
