### 2025-01-21 19f6448
Update figure
