+++ paper/paper.md
@@ -62,9 +62,7 @@ The [`eo_tides.utils`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_ti
 
 ## Modelling tides
 
-The [`eo_tides.model`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.model) module is powered by advanced tide modelling functionality from the `pyTMD` Python package [@pytmd].
-
-`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and earth tides. Tides are frequently decomposed into harmonic constants (or constituents) associated with the relative positions of the sun, moon and Earth. `pyTMD.io` contains routines for reading major constituent values from commonly available ocean tide models, and interpolating those values spatially. `pyTMD.astro` contains routines for computing the positions of celestial bodies for a given time. For ocean tides, `pyTMD` computes the longitudes of the sun (S), moon (H), lunar perigree (P), ascending lunar node (N) and solar perigree (PP). `pyTMD.arguments` combines astronomical coefficients with the "Doodson number" of each constituent, and adjusts the amplitude and phase of each constituent based on their modulations over the 18.6 year nodal period. Finally, `pyTMD.predict` uses results from those underlying functions to predict tidal values at a given location and time.
+The [`eo_tides.model`](https://geoscienceaustralia.github.io/eo-tides/api/#eo_tides.model) module is powered by advanced tide modelling functionality from the `pyTMD` Python package [@pytmd]. `pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and earth tides. Tides are frequently decomposed into harmonic constants (or constituents) associated with the relative positions of the sun, moon and Earth. `pyTMD.io` contains routines for reading major constituent values from commonly available ocean tide models, and interpolating those values spatially. `pyTMD.astro` contains routines for computing the positions of celestial bodies for a given time. For ocean tides, `pyTMD` computes the longitudes of the sun (S), moon (H), lunar perigree (P), ascending lunar node (N) and solar perigree (PP). `pyTMD.arguments` combines astronomical coefficients with the "Doodson number" of each constituent, and adjusts the amplitude and phase of each constituent based on their modulations over the 18.6 year nodal period. Finally, `pyTMD.predict` uses results from those underlying functions to predict tidal values at a given location and time.
 
 The `model_tides` function from `eo_tides.model` wraps `pyTMD` functionality to return tide predictions in a standardised `pandas.DataFrame` format, enabling integration with satellite EO data and parallelized processing for improved performance. Parallelisation in `eo-tides` is automatically optimised based on available workers and requested tide models and tide modelling locations. This parallelisation can significantly improve tide modelling performance, especially for large-scale analyses run on a multi-core machine (\autoref{tab:benchmark}). Additional functions like `model_phases` classify high, low or flow/ebb tides, critical for interpreting satellite-observed coastal processes like changing turbidity and ocean colour [@sent2025time].
 
@@ -92,7 +90,7 @@ Table: Comparison of the `tag_tides` and `pixel_tides` functions. \label{tab:tid
 
 ## Calculating tide statistics and satellite biases
 
[…]
