### 2025-01-21 cdac3ee
JOSS pre-review edits part 1 (#42)
* Fix citations

* First attempt at reducing word count

* Simplify

* Update generated PDF

* Small fixes

* Update generated PDF

* Update figure size

* Update generated PDF

* Simplify

* Update generated PDF

* Shift paragraph

* Update generated PDF

* Small fixes

* Update generated PDF

* Simplification

* Update generated PDF

* Minor edits

* Update generated PDF

* Replace image

* Update generated PDF

* Minor wording fixes

* Update generated PDF

* Updates

* Update generated PDF

* Fix spacing

* Update generated PDF

* Spacing

* Update genera
[…]
