### 2025-01-22 b221ece
JOSS pre-review edits part 2 (#43)
* Formatting fixes

* Update generated PDF

* Edits

* Update generated PDF

* Edits

* Update generated PDF

* Edits

* Update generated PDF

* Edits

* Update generated PDF

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
