### 2025-01-23 5b165fe
JOSS pre-review edits part 3 (#44)
* Tighten wording

* Update generated PDF

* Re-order table

* Update generated PDF

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
