### 2025-01-23 #7700 @robbibt
Hi @kthyng, I've been able to significantly reduce the length of the paper by approximately half: from **2675 to 1480 words.**

Would this be acceptable to move forward? I've made sure that the "summary" and "statement of need" are within the requested limit. The remainder is a brief summary of the key features of the package - I'd really love to keep this in the paper if at all possible. 🙂

Here are some suggestions for reviewers with expertise in Python packages designed for coastal analysis:

* Kilian Vos (kvos), author CoastSat Python package
* Sharon Fitzpatrick (2320sharon), author CoastSeg Python package
* Floris Calkoen (FlorisCalkoen), author coastpy Python package
* Freya Muir (fmemuir), author vegsat Python package
* Nicholas Pucino (npucino), author sandpyper Python package
