### 2025-01-27 #7700 @kthyng
@robbibt I appreciate the magnitude of words you've removed, but please do get down to about 1000 words. I'd recommend moving the more detailed sections about how to use the software into documentation.
