### 2025-01-27 #7700 @robbibt
Thanks @kthyng, and apologies for the hassle here: I was basing my submission on this previous JOSS paper I really liked which has a word count of around 3400 words, so I didn't know how strictly the word limit was applied: https://joss.theoj.org/papers/10.21105/joss.06683

I'll try and reduce it a little further. We do have more detailed content in our docs, but I was really hoping to keep at least a brief summary of the main features and some of our figures to help articulate how the different parts of the package fit together.
