### 2025-01-27 #7700 @kthyng
Ah, yeah we have started to enforce the word limit now, but that's confusing for people who have submitted before. Sorry about that. If you think it's really helpful to be in there you can go up to maybe 1200 words, but it may go up during review too.
