### 2025-01-28 0ee8a17
Additional shortening
