### 2025-01-28 970243b
Restore figure
