### 2025-01-29 afc08af
JOSS pre-review edits part 4 (#45)
* Tighten wording

* Update generated PDF

* Re-order table

* Additional shortening

* Update generated PDF

* Adjust formatting

* Update generated PDF

* Restore figure

* Update generated PDF

* Additional edits

* Update generated PDF

* Small edit

* Update generated PDF

* Add missing GESLA ref

* Update generated PDF

* Edits

* Update generated PDF

* Edits

* Update generated PDF

* Edits

* Update generated PDF

* 1 sentence is fine (#46)

fix

* Update generated PDF

* Combine paragraphs

* Update generated PDF

* Update paper.md

* Update generat
[…]
