### 2025-01-29 #7700 @robbibt
@kthyng @pdebuyl I have now cut it down to **1173** words. Thanks for your help so far, and please let me know if I can do anything else to help prepare it for review.
