+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.5.1 (upcoming)
+
+### New features
+
+- Added support for [Technical University of Denmark tide models](https://doi.org/10.11583/DTU.23828874) (DTU23)
+- Minor docs improvements, updates for new FES2022 data format
+
 ## v0.5.0
 
 ### New features
