### 2025-02-07 151c624
Test notebook outputs
