### 2025-02-10 5491eb8
Address #33 by adding `return_phase` param to `tag_tides` (#48)
* Addess #33 by adding return_phase

* Formatting

* Update generated PDF

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
