+++ docs/setup.md
@@ -41,14 +41,16 @@ Follow the guides below for some of the most commonly used global ocean tide mod
 
     3. Create a new directory inside your [tide model directory](#setting-up-a-tide-model-directory) called `EOT20/` to store the EOT20 model files.
 
-    4. Extract the `85762.zip` and then `ocean_tides.zip` into this new directory. You should end up with the following directory structure containing the extracted NetCDF files:
+    4. Extract the `85762.zip` and then `ocean_tides.zip` into this new directory.
 
-        ```
-        tide_models/EOT20/ocean_tides/
-           |- 2N2_ocean_eot20.nc
-           |- ...
-           |- T2_ocean_eot20.nc
-        ```
+    You should end up with the following directory structure containing the extracted NetCDF files:
+
+    ```
+    tide_models/EOT20/ocean_tides/
+        |- 2N2_ocean_eot20.nc
+        |- ...
+        |- T2_ocean_eot20.nc
+    ```
 
 ??? note "FES2022 Finite Element Solution tide models"
 
@@ -74,12 +76,12 @@ Follow the guides below for some of the most commonly used global ocean tide mod
 
     7. Navigate to `/auxiliary/tide_model/`, and download the contents of one of the following FTP directory:
 
-    - `fes2022b/ocean_tide_20241025/` (if using standard model data)
-    - `fes2022b/ocean_tide_extrapolated/` (if using extrapolated model data)
+        - `fes2022b/ocean_tide_20241025/` (if using standard model data)
+        - `fes2022b/ocean_tide_extrapolated/` (if using extrapolated model data)
 
-    !!! tip
+        !!! tip
 
-        The "extrapolated" version of FES models have been extended inland using a simple "nearest" extrapolation method to ensure data coverage across the entire coastal zone. This can be useful for ensuring you always return a modelled tide, but can also introduce uncertainty into your modelling (particularly in complex regions such as narrow peninsulas or inlets/embayments).
+            The "extrapolated" version of FES models have been extended inland using a simple "nearest" extrapolation method to ensure data coverage across the entire coastal zone. This can be useful for ensuring you always return a modelled tide, but can also introduce uncertainty into your modelling (particularly in complex regions such as narrow peninsulas or inlets/embayments).
 
     8. Create a new nested folder inside your [tide model directory](#setting-up-a-tide-model-directory) called either `fes2022b/ocean_tide_20241025/` (if using standard model data) or `fes2022b/ocean_tide_extrapolated/` (if using extrapolated model data) to store your dowloaded files.
 
@@ -87,19 +89,19 @@ Follow the guides below for some of the most commonly used global ocean tide mod
 
     You should end up with the following directory structure containing uncompressed `.nc`-format NetCDF files:
 
-        ```
-        tide_models/fes2022b/ocean_tide_20241025/
-           |- 2n2_fes2022.nc
-           |- ...
-           |- t2_fes2022.nc
-        ```
-        Or:
-        ```
-        tide_models/fes2022b/ocean_tide_extrapolated/
-           |- 2n2_fes2022.nc
-           |- ...
-           |- t2_fes2022.nc
-        ```
+    ```
+    tide_models/fes2022b/ocean_tide_20241025/
+        |- 2n2_fes2022.nc
+        |- ...
+        |- t2_fes2022.nc
+    ```
+    Or:
+    ```
[…]
