+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.6.0 (upcoming)
+## v0.6.0
 
 ### New features
 
