+++ .github/workflows/check-links.yml
@@ -1,0 +1,33 @@
+name: Check for broken links
+
+on:
+  workflow_dispatch: # Triggered manually
+
+jobs:
+  link-checker:
+    runs-on: ubuntu-latest
+    permissions:
+      issues: write
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+        with:
+          submodules: true
+
+      - name: Check links
+        id: check-links
+        uses: lycheeverse/lychee-action@v2
+        with:
+          fail: false
+          output: /tmp/broken-links-report.md
+          args: >
+            # --exclude-path ./docs/data/archive
+            ./**/*.md ./**/*.rst ./**/*.ipynb ./**/*.yaml ./**/*.yml
+
+      - name: Create issue
+        if: steps.check-links.outputs.exit_code != 0
+        uses: peter-evans/create-issue-from-file@v5
+        with:
+          title: Fix broken links
+          content-filepath: /tmp/broken-links-report.md
+          labels: link-checker
