### 2025-02-18 #7786 @pdebuyl
@fmemuir @daniloceano please check the information text at the top of this page and create your checklist (this is not done by the editorial bot so the checklist "belongs" to your GitHub user).

As the review takes place into this discussion page, feel free to provide a partial review at some point, to ask clarifications to the author here, or to ask me about the review process.

Some reviewers also open issues at the repository under review. This is fine for bug reports or requests about the documentation, for instance. In that case, please report the resulting bugfix or other result from the discussion here so that I can get proper notification.

@robbibt : for information, both reviewers accepted to review but have a busy schedule until end February.
