### 2025-02-20 8ba4eb3
Fix bbox clipping bug #50
