+++ docs/changelog.md
@@ -4,7 +4,7 @@
 
 ### Bug fixes
 
-- Fixed bug causing tide model clippin with `clip_tides` to fail for bounding boxes west of the prime meridian ([#50](https://github.com/GeoscienceAustralia/eo-tides/issues/50))
+- Fixed bug causing tide model clipping with `clip_tides` to fail for bounding boxes completely west of the prime meridian ([#50](https://github.com/GeoscienceAustralia/eo-tides/issues/50))
 
 ## v0.6.0
 
