### 2025-02-20 616f38f
Add test of global clipping options
