+++ tests/conftest.py
@@ -102,7 +102,9 @@ def satellite_ds(satellite_ds_load):
     return deepcopy(satellite_ds_load)
 
 
-@pytest.fixture(scope="session")
+# Run once per session to generate symethic HAMTIDE11 files; autouse=True
+# allows this to run without being specifically called in tests
+@pytest.fixture(scope="session", autouse=True)
 def create_synthetic_model(base_dir="data/tide_models_synthetic"):
     """
     Generates and exports synthetic HAMTIDE11 model data
