### 2025-02-20 624f70a
Fix bounding box clipping bug #50 (#55)
* Fix bbox clipping bug #50

* Update lockfile

* Updates for compat with pytmd==2.2.2

* Fix time

* New version 0.6.1

* Fix wording

* Add test of global clipping options

* Autouse synethic hamtide fixture

* Fix output path
