+++ mkdocs.yml
@@ -44,7 +44,7 @@ plugins:
             # docstring_options:
             #   ignore_init_summary: true
             # docstring_section_style: list
-            # filters: ["!^_"]
+            filters: ["!^_"]
             # heading_level: 1
             # inherited_members: true
             # merge_init_into_class: true
@@ -52,7 +52,6 @@ plugins:
             separate_signature: true
             line_length: 60
             show_root_heading: true
-            show_private_members: false
             # show_category_heading: true
             # show_root_full_path: true
             # show_root_members_full_path: true
