### 2025-02-23 a80841b
Change on-the-fly cropping to "auto"
