### 2025-02-24 014c5ff
Simplify package build and publish
