+++ .github/actions/setup-python-env/action.yml
@@ -13,7 +13,7 @@ inputs:
   uv-version:
     description: "uv version to use"
     required: true
-    default: "0.5.0"
+    default: "0.6.2"
 
 runs:
   using: "composite"
