+++ .github/workflows/on-release-main.yml
@@ -45,13 +45,12 @@ jobs:
           name: pyproject-toml
 
       - name: Build package
-        run: uvx --from build pyproject-build --installer uv
+        run: uv build
 
       - name: Publish package
-        run: uvx twine upload dist/*
+        run: uv publish
         env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
+          UV_PUBLISH_TOKEN: ${{ secrets.PYPI_TOKEN }}
 
   deploy-docs:
     needs: publish
