+++ .github/workflows/on-release-main.yml
@@ -2,7 +2,8 @@ name: release-main
 
 on:
   release:
-    types: [published, edited]
+    types: [published]
+  push:
     branches: [main]
 
 jobs:
@@ -16,12 +17,22 @@ jobs:
         run: echo tag=${GITHUB_REF#refs/*/} >> $GITHUB_OUTPUT
         if: ${{ github.event_name == 'release' }}
 
-      - name: Update project version
+      # For full releases, use tag for version name
+      - name: Set release version
         run: |
           sed -i "s/^version = \".*\"/version = \"$RELEASE_VERSION\"/" pyproject.toml
         env:
           RELEASE_VERSION: ${{ steps.vars.outputs.tag }}
         if: ${{ github.event_name == 'release' }}
+
+      # For merges to main, use ".devN" suffix based on commits since tag
+      - name: Set dev version
+        run: |
+          LATEST_TAG=$(git describe --tags --abbrev=0)
+          DEV_NUM=$(git rev-list $LATEST_TAG..HEAD --count)
+          DEV_VERSION="$LATEST_TAG.dev$DEV_NUM"
+          sed -i "s/^version = \".*\"/version = \"$DEV_VERSION\"/" pyproject.toml
+        if: ${{ github.event_name == 'push' }}
 
       - name: Upload updated pyproject.toml
         uses: actions/upload-artifact@v4
