### 2025-02-24 66c1668
Simplify package build and publish (#59)
* Simplify package build and publish

* Add dev release workflow
