+++ .github/workflows/on-release-main.yml
@@ -11,6 +11,8 @@ jobs:
     runs-on: ubuntu-24.04
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0  # Need full history for tags
 
       - name: Export tag
         id: vars
