### 2025-03-08 #7786 @fmemuir
@pdebuyl Quick question: is there a difference between the Statement Of Need required for the software paper and the one listed under **Documentation** in the reviewer checklist? Should the docs at [https://geoscienceaustralia.github.io/eo-tides/](https://geoscienceaustralia.github.io/eo-tides/) contain its own Statement Of Need, or is the one in the paper sufficient?
