### 2025-03-09 #7786 @pdebuyl
Hello @fmemuir the statement of need is a mandatory section in the article. It also needs to be present in some form in the documentation of the software. Documentation is expected in the form of a "user guide" (which in the most basic form can be an extended README) and of some sort of API documentation.
