### 2025-03-10 #7786 @daniloceano
@robbibt Thank you for the package! From what I’ve seen so far, I’m sure it’s a great contribution and will be very useful for other researchers. I’m currently downloading the model data, and as I test the package, I’ll share my comments, suggestions, and feedback in this thread.
