### 2025-03-10 #7786 @daniloceano
First of all, @robbibt, in the documentation section "Downloading tide model data," I believe it might be helpful to clarify that what is being downloaded are the harmonic components rather than the model outputs themselves. Apologies for being picky about this detail, but I think it could help prevent potential confusion among users. I personally had this question when I reached this part and had to refer back to the paper for clarification.
