### 2025-03-10 #7786 @robbibt
> First of all, @robbibt, in the documentation section "Downloading tide model data," I believe it might be helpful to clarify that what is being downloaded are the harmonic components rather than the model outputs themselves. Apologies for being picky about this detail, but I think it could help prevent potential confusion among users. I personally had this question when I reached this part and had to refer back to the paper for clarification.

Not picky at all - it's great feedback and definitely a common point of confusion. I'll add some extra info to this part of the docs today and tag you as a reviewer on the PR if that's OK.
