### 2025-03-11 #7786 @robbibt
@daniloceano @fmemuir I invited you both to `eo-tides` with "triage" permissions so that you can review any PRs I make - if you didn't get the invite this should work: https://github.com/GeoscienceAustralia/eo-tides/invitations
