### 2025-03-11 cdbb01d
Add background context for tide modelling files (#60)
