### 2025-03-11 #7786 @daniloceano
![Image](https://github.com/user-attachments/assets/485e9dc1-6b9c-42b8-8db9-89ae40dc927f)

@robbibt In this section, it might be helpful to clarify that the "directory" variable should point to the full or relative path of the tide_models directory. While the current example is already good, in my experience, it's always best to assume that some users may not be entirely sure what to do. Providing a more explicit example, such as a full path, could make it clearer.

Additionally, adding a comment in the code itself—something like:

`directory="tide_models/"  # Fill with full or relative path to tide_models directory`

— might further improve usability and prevent confusion. Just a small suggestion to enhance clarity! 😊

But it completely is up to up, it is just a suggestion though.
