### 2025-03-11 GeoscienceAustralia/eo-tides#62 @daniloceano
Bug report: clipping model data
@robbibt I encountered an issue while trying to clip the tide models to the Brazilian coastline using clip_models(). Initially, I ran:

```
from eo_tides.utils import clip_models

clip_models(
    input_directory="tide_models/",
    output_directory="tide_models_Brazil/",
    bbox=(-55, -35, -23, 6),
)

```
At first, I got:

> Traceback (most recent call last):
>   File "/home/daniloceano/Documents/Programs_and_scripts/eo-tides/test_eo-tides.py", line 3, in <module>
>     clip_models(
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/eo_tides/utils.py", line 372, in clip_models
>     nc = xr.open_mfdataset(input_directory / file)
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/xarray/backends/api.py", line 1077, in open_mfdataset
>     datasets = [open_(p, **open_kwargs) for p in paths]
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/xarray/backends/api.py", line 1077, in <listcomp>
>     datasets = [open_(p, **open_kwargs) for p in paths]
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/xarray/backends/api.py", line 594, in open_dataset
>     ds = _dataset_from_backend_dataset(
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/xarray/backends/api.py", line 370, in _dataset_from_backend_dataset
>     ds = _chunk_ds(
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/xarray/backends/api.py", line 318, in _chunk_ds
>     chunkmanager = guess_chunkmanager(chunked_array_type)
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/xarray/namedarray/parallelcompat.py", line 117, in guess_chunkmanager
>     raise ValueError(
> ValueError: unrecognized chunk manager dask - must be one of: []

which I assumed was due to a missing Dask installation. After installing Dask via pip, I tried again and encountered a new error:

> Traceback (most recent call last):
>   File "/home/daniloceano/Documents/Programs_and_scripts/eo-tides/test_eo-tides.py", line 3, in <module>
>     clip_models(
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/eo_tides/utils.py", line 407, in clip_models
>     nc_clipped = _clip_model_file(nc, bbox, xdim="lon", ydim="lat", ycoord="lat", xcoord="lon")
>   File "/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.9/site-packages/eo_tides/utils.py", line 275, in _clip_model_file
>     return nc_clipped
> UnboundLocalError: local variable 'nc_clipped' referenced before assignment

Interestingly, when I use the example provided in the documentation, everything works fine:

![Image](https://github.com/user-attachments/assets/2bf4b50b-50f5-488b-b2b1-9aa644c32ba5)

 It seems that the error might be related to how the function handles certain inputs.

Would it be possible to provide a more informative error message in cases where the clipping fails? That could help users diagnose the issue more easily. Thanks!

https://github.com/openjournals/joss-reviews/issues/7786
