### 2025-03-11 GeoscienceAustralia/eo-tides#62 @robbibt
Thanks @daniloceano! I've had some issues with clipping model files in certain areas recently - I'll investigate this right away and either add better error handling or fix the issue itself.

As a workaround, hopefully everything will work on the default unclipped model files!
