### 2025-03-11 GeoscienceAustralia/eo-tides#62 @robbibt
@daniloceano What tide model(s) did you download/are you attempting to clip here?
