### 2025-03-11 GeoscienceAustralia/eo-tides#62 @daniloceano
Oh sorry, I should have mentioned that before. I am using the EOT20 model
