### 2025-03-12 GeoscienceAustralia/eo-tides#62 @robbibt
Hey @daniloceano - unfortunately I can't seem to reproduce this on my end, e.g:

![Image](https://github.com/user-attachments/assets/7b9588e5-1ae8-4305-b9c8-f2bce343a992)

Plotting the outputs look OK:

![Image](https://github.com/user-attachments/assets/114869c2-de8c-498d-990e-1a8c686de2dd)

I suspect this is a package dependency issue - can you please run `pip list` in your environment and post the results? Particularly, am interested in the versions of `eo-tides`, `xarray`, `numpy`, `pandas`, `Dask` and `pyTMD`.
