### 2025-03-12 e7b81ad
Expand clipping test to include synethic EOT20
