### 2025-03-12 GeoscienceAustralia/eo-tides#62 @robbibt
Actually, one thing I can see is `python3.9` in the logs - the package technically isn't compatible with Python 3.9 (it only supports 3.10 and above). I'm not sure if this is actually related to the issue above (I would have thought this would have blocked it from installing at all), but if you have a way to install Python 3.10 it would be interesting to see if you encounter the same issue!
