### 2025-03-12 GeoscienceAustralia/eo-tides#62 @daniloceano
**@robbibt** I continued following the user guide instructions:

![Image](https://github.com/user-attachments/assets/457fa797-50db-4c22-a759-de96be1b7903)

 but adapted them to my region of interest

I have two suggestions:
1. It might be helpful to explicitly include the module import for `model_tides` in the code block to ensure clarity.
2. I'm encountering a warning when running the following code:

```python
from eo_tides.model import model_tides
import pandas as pd

model_tides(
    x=-40,
    y=-20,
    time=pd.date_range("2022-01-01", "2022-01-04", freq="1D"),
    directory="tide_models_Brazil/"
)
```

/home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.10/site-packages/eo_tides/model.py:125: UserWarning: On-the-fly cropping is not compatible with the provided clipped model files; running with `crop=False`.
  warnings.warn(

Is this warning expected?
