### 2025-03-12 GeoscienceAustralia/eo-tides#62 @robbibt
> /home/daniloceano/miniconda3/envs/eo-tides-env/lib/python3.10/site-packages/eo_tides/model.py:125: UserWarning: On-the-fly cropping is not compatible with the provided clipped model files; running with `crop=False`. warnings.warn(
>
> Is this warning expected?

Yes: `eo-tides` applies two types of clipping/cropping to improve performance: the manual approach that you've been running to actually reduce the size of the model files, and an "on-the-fly" method done by the underlying `pyTMD` library. Usually it's fine to combine them, but unfortunately this doesn't work for study areas that are entirely restricted to the western hemisphere. There's more detail here:

- https://github.com/GeoscienceAustralia/eo-tides/pull/58
- https://github.com/GeoscienceAustralia/eo-tides/issues/50#issuecomment-2676630589

I think I'll improve the warning: would this help?

```
On-the-fly cropping is not compatible with the provided model files; running with `crop=False`. This will not affect your results but may lead to a minor slowdown. This can occur when analysing clipped model files restricted to the western hemisphere. To suppress this warning, manually set `crop=False`.
```
