### 2025-03-12 ad85fb4
Use tabs for file paths, update clipping warning
