+++ mkdocs.yml
@@ -117,6 +117,8 @@ markdown_extensions:
       generic: true
   - admonition
   - pymdownx.details
+  - pymdownx.tabbed:
+      alternate_style: true
   - pymdownx.superfences
   - pymdownx.highlight:
       anchor_linenums: true
