### 2025-03-12 b8e11ac
Update directory path documentation, improve warnings (#65)
* Add background context for tide modelling files

* Add full path examples

* Use tabs for file paths, update clipping warning
