### 2025-03-13 GeoscienceAustralia/eo-tides#62 @robbibt
> **[@robbibt](https://github.com/robbibt)** After reinstalling my environment with Python 3.10, everything is working fine! Thanks for the help.
>
> Would it be possible to add installation instructions in the documentation, specifying how to set up the environment using `conda` or a virtual environment? This could ensure that users start with the correct dependencies.
>
> At the very least, making the Python 3.10 requirement more explicit in the installation guide could help prevent others from running into the same issue I did. (Apologies if this is already mentioned somewhere and I missed it!)
>
> Thanks again for the support! 😊

More great feedback, thanks! Will work on this - ideally I'd love to use a `uv` approach instead of `conda` because it's a really powerful and easy way to manage dependencies... any change you could try running this on your end and see if it sucessfully installs the package for you?

```
# Install `uv`
pip install uv

# Create a new virtual environment
uv venv

# Activate virtual environment
source .venv/bin/activate

# Install `eo-tides` with all required dependencies
uv pip install eo-tides
```
