### 2025-03-13 GeoscienceAustralia/eo-tides#62 @daniloceano
> I think I'll improve the warning: would this help?
>
> > On-the-fly cropping is not compatible with the provided model files; running with `crop=False`. This will not affect your results but may lead to a minor slowdown. This can occur when analysing clipped model files restricted to the western hemisphere. To suppress this warning, manually set `crop=False`.

That's perfect.
