### 2025-03-13 GeoscienceAustralia/eo-tides#62 @daniloceano
**@robbibt** I ran into an issue when trying to clip the tide models using `eo-tides` installed via `uv`. I got the following error:

`ImportError: chunk manager 'dask' is not available. Please make sure 'dask' is installed and importable.`

After manually installing `dask` with:

```bash
uv pip install dask
```

The issue was resolved, and everything worked fine.

I’m wondering if dask might not be listed as an explicit dependency in eo-tides, which could explain why it wasn’t installed automatically. I don’t know enough about uv's dependency resolution to say for sure, but maybe it’s something worth checking.

Just wanted to flag this in case it affects other users.
