### 2025-03-13 #7786 @daniloceano
**@robbibt** I’ve finished testing everything related to the **Modeling Tides** section of the documentation, and everything is working great! 🚀

I ran tests for my region of interest and also tried to break the code in different ways by tweaking some parameters to see how it would behave. The only minor thing I noticed is that the example for plotting tide phase periods only works if `model_phases` is used with the argument `output_format="long"`. But this is a very minor detail.

Overall, fantastic work—everything is running smoothly!
