### 2025-03-13 GeoscienceAustralia/eo-tides#62 @robbibt
> > ```
> >
> > # Create a new virtual environment
> > uv venv
> >
> > # Activate virtual environment
> > source .venv/bin/activate
> >
> > # Install `eo-tides` with all required dependencies
> > uv pip install eo-tides```
> > ```
>
> **[@robbibt](https://github.com/robbibt)** I tested it, and everything seems to be working fine! I wasn’t familiar with `uv`, but it looks like a great tool. I’ll let you know if I run into any issues in the future.
>
> Would it make sense to add a note in the documentation about using `uv` for setting up the environment? Maybe also mention that `conda` and other package managers might work, but haven’t been explicitly tested? Just a suggestion—of course, only if it makes sense to you.

Yep, will update the install docs to add a section on `uv`, I think that would be a handy addition!
