### 2025-03-14 7c6f8b1
JOSS review updates: Add dask dependency, add tip about using print for dataframes (#68)
* Add dask depend, use print

* Ignore dask in deptry

* Revert using print, add tip instead

* Wording

* Wording
