### 2025-03-14 a0c4e44
Update install docs (#69)
