### 2025-03-18 #7786 @robbibt
Hey @daniloceano, `x` and `y` are generic because you can use the `crs` param to provide coordinates in any other non lat/lon projection system too. 🙂
