### 2025-03-19 0cb3410
Ensure dtype of ensemble modelled tides matches dtype of input modelled tides (#70)
* Ensure dtype matches input dtype in ensemble modelling

* Add ensemble dtype test

* Update comments
