### 2025-03-20 8f7cb81
Prevent tests from running twice on subsequent changes (#72)
* Prevent tests from running twice on subsequent changes

* Update main.yml

* Fix format
