### 2025-03-25 GeoscienceAustralia/eo-tides#74 @daniloceano
Issue: No image output when running "Load satellite data using odc-stac" example
I tried running the example in the **"Load satellite data using odc-stac"** section. The code executed without any apparent errors, and I even added `plt.show()` at the end, but no image was displayed.

The following warnings appeared:

/home/daniloceano/Documents/Programs_and_scripts/eo-tides/.venv/lib/python3.12/site-packages/odc/geo/_rgba.py:55: RuntimeWarning: invalid value encountered in cast return x.astype("uint8") /home/daniloceano/Documents/Programs_and_scripts/eo-tides/.venv/lib/python3.12/site-packages/rasterio/warp.py:387: NotGeoreferencedWarning: Dataset has no geotransform, gcps, or rpcs. The identity matrix will be returned. dest = _reproject(

I tested this both in VS Code and directly in the terminal, with the same result — no image output.

https://github.com/openjournals/joss-reviews/issues/7786
