### 2025-03-25 GeoscienceAustralia/eo-tides#74 @robbibt
I suspect this is related to #73 - would be great to see if the same issue persists on a clean environment installed via
```
uv pip install eo-tides[notebooks]
```
