+++ pyproject.toml
@@ -36,7 +36,7 @@ classifiers = [
 requires-python = ">=3.10,<4.0"
 dependencies = [
     "colorama>=0.4.3",
-    "dask[complete]>=2023.1.0",
+    "dask[complete]>=2023.1.0,!=2025.3.0",
     "geopandas>=0.10.0",
     "matplotlib>=3.8.0",
     "numpy>=1.26.0",
