### 2025-03-31 GeoscienceAustralia/eo-tides#74 @robbibt
I've just closed #73 as I believe the installation instructions should now work for creating a working environment (the problem was an external Dask issue that caused `uv pip install eo-tides[notebooks]` to produce an invalid environment). Fingers re-creating your virtual environment on the latest version fixes this issue too! 🤞
