### 2025-03-31 GeoscienceAustralia/eo-tides#62 @robbibt
Closing this as we now correctly include Dask as a dependency, and the installation docs have been updated to cover creating a consistent environment with `uv`. Thanks so much for this feedback, it's been super useful!
