### 2025-04-01 GeoscienceAustralia/eo-tides#74 @daniloceano
I am still am not able to display a image when running the example. Please see bellow the code along the terminal output:

```
import odc.stac
import pystac_client
import matplotlib.pyplot as plt

directory = "tide_models_clipped/"

# Connect to STAC catalog
print("Connecting to STAC catalog...")
catalog = pystac_client.Client.open("https://explorer.dea.ga.gov.au/stac")
print("Connected to STAC catalog.")

# Set cloud access defaults
print("Setting cloud access defaults...")
odc.stac.configure_rio(
    cloud_defaults=True,
    aws={"aws_unsigned": True},
)
print("Cloud access defaults set.")

# Build a query and search the STAC catalog for all matching items
print("Querying STAC catalog...")
bbox = [122.12, -18.25, 122.43, -17.93]
query = catalog.search(
    bbox=bbox,
    collections=["ga_ls8c_ard_3", "ga_ls9c_ard_3"],
    datetime="2024-01-01/2024-12-31",
    filter = "eo:cloud_cover < 5"  # Filter to images with <5% cloud
)
print("Query complete.")

# Load data into xarray format
print("Loading data into xarray format...")
ds = odc.stac.load(
    items=list(query.items()),
    bands=["nbart_red", "nbart_green", "nbart_blue"],
    crs="utm",
    resolution=30,
    groupby="solar_day",
    bbox=bbox,
    fail_on_error=False,
    chunks={},
)
print("Data loaded into xarray format.")

# Plot the first image
print("Plotting first image...")
ds.isel(time=0).odc.explore(vmin=50, vmax=3000)
# plt.show()
```

> Connecting to STAC catalog...
Connected to STAC catalog.
Setting cloud access defaults...
Cloud access defaults set.
Querying STAC catalog...
Query complete.
Loading data into xarray format...
Data loaded into xarray format.
Plotting first image...
/home/daniloceano/Documents/Programs_and_scripts/eo-tides/.venv/lib/python3.10/site-packages/odc/geo/_rgba.py:55: RuntimeWarning: invalid value encountered in cast
  return x.astype("uint8")
/home/daniloceano/Documents/Programs_and_scripts/eo-tides/.venv/lib/python3.10/site-packages/rasterio/warp.py:387: NotGeoreferencedWarning: Dataset has no geotransform, gcps, or rpcs. The identity matrix will be returned.
  dest = _reproject(

Just in case , this is how I have clipped the data:

```
from eo_tides.utils import clip_models
import pandas as pd

clip_models(
    input_directory="tide_models/",
    output_directory="tide_models_clipped/",
    bbox=(113.3, -43.6, 153.6, -10.7),
)
```
