### 2025-04-01 GeoscienceAustralia/eo-tides#74 @robbibt
@kirill888 Would you expect `odc-geo` tools like `.odc.explore()` (that create and add data to an interactive folium map) to work when running code from a terminal? I'm just wondering if they only work correctly via Jupyter Lab.

@daniloceano You can produce a non-interactive map via something like `ds.isel(time=0).odc.to_rgba().plt.imshow()`. This isn't strictly an `eo-tides` issue, so if `ds` does contain valid xarray data the rest of the tide modelling code should all work fine (the other warnings you posted are also unrelated and come from `odc-stac`).
