### 2025-04-02 GeoscienceAustralia/eo-tides#74 @Kirill888
@robbibt that's not what's failing, to_rgba is failing as numpy on that machine doesn't recognise uint8 as a valid data type
