### 2025-04-02 GeoscienceAustralia/eo-tides#74 @robbibt
> [@robbibt](https://github.com/robbibt) that's not what's failing, to_rgba is failing as numpy on that machine doesn't recognise uint8 as a valid data type

The `RuntimeWarning: invalid value encountered in cast return x.astype("uint8")` warning? That's coming directly after an `odc.stac.load()`, so does this suggest an `odc.stac` issue, e.g. data types that aren't compatible with `odc-geo` are being returned?

Confusingly, if I run the code as a Jupyter Notebook from the same environment, the data plots fine for me without any warning:
```
curl -LsSf https://astral.sh/uv/install.sh | sh

uv venv --python 3.13
source .venv/bin/activate

uv pip install eo-tides[notebooks]

uv run --with jupyter jupyter lab Satellite_data.ipynb
```

![Image](https://github.com/user-attachments/assets/d6d68297-9937-4f11-b7b8-44b64447ff78)
```
Package                   Version
------------------------- -----------
affine                    2.4.0
annotated-types           0.7.0
attrs                     25.3.0
bokeh                     3.7.2
branca                    0.8.1
cachetools                5.5.2
certifi                   2025.1.31
cftime                    1.6.4.post1
charset-normalizer        3.4.1
click                     8.1.8
click-plugins             1.1.1
cligj                     0.7.2
cloudpickle               3.1.1
colorama                  0.4.6
contourpy                 1.3.1
cycler                    0.12.1
dask                      2025.2.0
distributed               2025.2.0
eo-tides                  0.6.4
folium                    0.19.5
fonttools                 4.56.0
fsspec                    2025.3.1
geopandas                 1.0.1
idna                      3.10
imagecodecs               2025.3.30
jinja2                    3.1.6
joblib                    1.4.2
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
kiwisolver                1.4.8
locket                    1.0.0
lxml                      5.3.1
lz4                       4.4.3
markupsafe                3.0.2
matplotlib                3.10.1
msgpack                   1.1.0
narwhals                  1.32.0
netcdf4                   1.7.2
numpy                     2.2.4
odc-geo                   0.4.10
odc-stac                  0.3.11
packaging                 24.2
pandas                    2.2.3
partd                     1.4.2
pillow                    11.1.0
planetary-computer        1.0.0
psutil                    7.0.0
pyarrow                   19.0.1
pydantic                  2.11.1
pydantic-core             2.33.0
pyogrio                   0.10.0
pyparsing                 3.2.3
pyproj                    3.7.1
pystac                    1.12.2
pystac-client             0.8.6
python-dateutil           2.9.0.post0
python-dotenv             1.1.0
pytmd                     2.2.2
pytz                      2025.2
pyyaml                    6.0.2
rasterio                  1.4.3
referencing               0.36.2
requests                  2.32.3
rpds-py                   0.24.0
scikit-learn              1.6.1
scipy                     1.15.2
setuptools                78.1.0
setuptools-scm            8.2.0
shapely                   2.0.7
six                       1.17.0
sortedcontainers          2.4.0
tblib                     3.0.0
threadpoolctl             3.6.0
tifffile                  2025.3.30
timescale                 0.0.8
toolz                     1.0.0
tornado                   6.4.2
tqdm                      4.67.1
typing-extensions         4.13.0
typing-inspection         0.4.0
tzdata                    2025.2
urllib3                   2.3.0
xarray                    2025.3.0
xyzservices               2025.1.0
zict                      3.0.0
```
