### 2025-04-02 GeoscienceAustralia/eo-tides#74 @Kirill888
I see so just a warning, I read it wrong. Instead of pltshow try adding display from ipython
