### 2025-04-02 GeoscienceAustralia/eo-tides#74 @robbibt
Ipython is related to Jupyter Notebooks and Jupyer Lab - so if "display" is required then it's possible these interactive features won't work outside of Jupyter?
