### 2025-04-02 GeoscienceAustralia/eo-tides#74 @robbibt
OK, thanks Kirill - I didn't really understand how these interactive maps actually worked. For my understanding, is it basically this issue here? https://github.com/python-visualization/folium/issues/946

When you say VScode extensions, do you mean something that can run Jupyter Notebooks, or some Ipython extension?
