### 2025-04-02 GeoscienceAustralia/eo-tides#74 @Kirill888
I guess this one: https://marketplace.visualstudio.com/items?itemName=ms-toolsai.jupyter-renderers but it probably comes as a pack of related extensions https://marketplace.visualstudio.com/items?itemName=ms-toolsai.jupyter
