### 2025-04-02 GeoscienceAustralia/eo-tides#74 @robbibt
Awesome, thanks!

@daniloceano I think this means you probably need to run the code as Jupyter Notebooks in VScode in order to get working interactive plots. They're not critical to the examples though, so shouldn't affect any of the key `eo-tides` tide modelling functionality.
