### 2025-04-03 GeoscienceAustralia/eo-tides#74 @daniloceano
**@robbibt** Using Jupyter worked, and I was able to see the interactive plots. However, I think it might be useful to clarify in the documentation that the image will only be displayed if using Jupyter (or something that can render HTML images). I was running everything directly from the terminal, and users who aren't familiar with the `odc` package might encounter the same difficulty.

Additionally, to get the virtual environment working in Jupyter, I had to do the following steps:

1) Install the interactive kernel:
  ```bash
   uv pip install ipykernel
  ```

2) Register the Jupyter kernel:

  ```
    python -m ipykernel install --user --name eo-tides --display-name "Python (eo-tides)"
  ```

I’m not a heavy Jupyter user (I actually don’t prefer it — I usually stick to the terminal or, at most, VS Code), so these steps weren’t immediately obvious to me.
