### 2025-04-03 #7786 @daniloceano
**@robbibt** I ran the **"Load satellite data using odc-stac"** example, and it worked fine! However, I was wondering how this could be adapted for other regions around the globe. I did some tests for regions like Xangai and others, but I wasn’t able to reproduce the images.

Here's the modified code I tried for Xangai:

```python
import odc.stac
import pystac_client
import matplotlib.pyplot as plt

# Connect to STAC catalog
print("Connecting to STAC catalog...")
catalog = pystac_client.Client.open("https://explorer.dea.ga.gov.au/stac")
print("Connected to STAC catalog.")

# Set cloud access defaults
print("Setting cloud access defaults...")
odc.stac.configure_rio(
    cloud_defaults=True,
    aws={"aws_unsigned": True},
)
print("Cloud access defaults set.")

# Build a query and search the STAC catalog for all matching items
print("Querying STAC catalog...")

bbox = [121.5, 29.5, 122.5, 30.5]  # Xangai

query = catalog.search(
    bbox=bbox,
    collections=["ga_ls8c_ard_3", "ga_ls9c_ard_3"],
    datetime="2024-01-01/2024-12-31",
    filter="eo:cloud_cover < 5"  # Filter to images with <5% cloud
)
print("Query complete.")

# Load data into xarray format
print("Loading data into xarray format...")
ds = odc.stac.load(
    items=list(query.items()),
    bands=["nbart_red", "nbart_green", "nbart_blue"],
    crs="utm",
    resolution=30,
    groupby="solar_day",
    bbox=bbox,
    fail_on_error=False,
    chunks={},
)
print("Data loaded into xarray format.")

# Plot the first image
print("Plotting first image...")
ds.isel(time=0).odc.explore(vmin=50, vmax=3000)
# plt.show()
```

The error I get is the following:

> Connecting to STAC catalog...
> Connected to STAC catalog.
> Setting cloud access defaults...
> Cloud access defaults set.
> Querying STAC catalog...
> Query complete.
> Loading data into xarray format...

> ---------------------------------------------------------------------------
> ValueError                                Traceback (most recent call last)
> Cell In[18], line 37
>      35 # Load data into xarray format
>      36 print("Loading data into xarray format...")
> ---> 37 ds = odc.stac.load(
>      38     items=list(query.items()),
>      39     bands=["nbart_red", "nbart_green", "nbart_blue"],
>      40     crs="utm",
>      41     resolution=30,
>      42     groupby="solar_day",
>      43     bbox=bbox,
>      44     fail_on_error=False,
>      45     chunks={},
>      46 )
>      47 print("Data loaded into xarray format.")
>      49 # Plot the first image
>

> File ~/Documents/Programs_and_scripts/eo-tides/.venv/lib/python3.10/site-packages/odc/stac/_stac_load.py:389, in load(items, bands, groupby, resampling, dtype, chunks, pool, crs, resolution, anchor, geobox, bbox, lon, lat, x, y, like, geopolygon, intersects, progress, fail_on_error, stac_cfg, patch_url, preserve_original_order, driver, **kw)
>     385 debug = kw.get("debug", False)
>     387 # Check we have all the bands of interest
>     388 # will raise ValueError if no such band[/alias](http://localhost:8888/alias)
> --> 389 collection = _collection(_parsed)
>     390 bands_to_load = collection.resolve_bands(bands)
>     391 bands = list(bands_to_load)
>
> File ~/Documents/Programs_and_scripts/eo-tides/.venv/lib/python3.10/site-packages/odc/stac/_stac_load.py:55, in _collection(items)
>      53 for item in items:
>      54     return item.collection
> ---> 55 raise ValueError("Can't load empty sequence")
>
> ValueError: Can't load empty sequence

While I understand this might be more related to familiarity with pystac and odc (which I wasn't familiar with previously), I think it could be helpful to provide some guidance for users outside of Australia on how to adapt these examples for other regions of interest. Since some functionalities in the eo-tides package are closely tied to these packages, offering at least a tip or two could assist users in reproducing examples for their own areas.
