### 2025-04-03 GeoscienceAustralia/eo-tides#74 @robbibt
Yep, good point. The examples here are primarily intended to be run in Jupyter Notebook format, but I can definitely add some additional notes to highlight anything that won't work on other platforms.

Thanks for all the feedback here - I realise it took a lot of effect to get to this point! (we use Jupyter Notebooks heavily at Geoscience Aus, so I'm not as familiar with how the code works outside of that)
