### 2025-04-03 #7786 @robbibt
> **[@robbibt](https://github.com/robbibt)** I ran the **"Load satellite data using odc-stac"** example, and it worked fine! However, I was wondering how this could be adapted for other regions around the globe. I did some tests for regions like Xangai and others, but I wasn’t able to reproduce the images.
> While I understand this might be more related to familiarity with pystac and odc (which I wasn't familiar with previously), I think it could be helpful to provide some guidance for users outside of Australia on how to adapt these examples for other regions of interest. Since some functionalities in the eo-tides package are closely tied to these packages, offering at least a tip or two could assist users in reproducing examples for their own areas.

Yeah, this is a fair point. The reason it doesn't work is because the `odc-stac` examples currently load data from several different STAC APIs: Microsoft Planetary Computer which has global coverage, and Digital Earth Australia which only has coverage for Australia.

I originally did this to try and demonstrate that `eo-tides` can be applied generically to any source of freely available satellite imagery, and isn't tied to one platform. But given we aim to support global use cases, it probably makes more sense to go all-on on a platform that has global coverage.

I'll see if I can easily update the examples to switch entirely to Microsoft Planetary Computer, and then you can see if they run OK for your other regions of interest?
