### 2025-04-04 GeoscienceAustralia/eo-tides#79 @robbibt
Feature suggestion: Update all `odc-stac` examples to use Microsoft Planetary Computer only
**Describe your idea or suggestion**
The current examples use a combination of the Digital Earth Australia and Microsoft Planetary Computer STAC API endpoints. This works for the provided examples, but the DEA API is restricted to Australia so doesn't work globally.

To ensure all examples work anywhere in the world, we could instead use Microsoft Planetary Computer in all examples.

More detail:
https://github.com/openjournals/joss-reviews/issues/7786#issuecomment-2776921111
