### 2025-04-04 #7786 @daniloceano
@robbibt That sounds like a great idea — the original intent behind the example is excellent, and adapting it for other regions using global coverage sources like Microsoft Planetary Computer would be amazing. I'm sure that would help many future users! Personally, I already see myself as a user of this package, and I’ve actually recommended it to several colleagues here in my department.
