### 2025-04-04 #7786 @robbibt
> Sorry @robbibt I've been a bit swamped with conference prep  - could you please send the collaboration invite again?

No worries, hope this works! https://github.com/GeoscienceAustralia/eo-tides/invitations
