### 2025-04-04 #7786 @fmemuir
> [@robbibt](https://github.com/robbibt) In this section, it might be helpful to clarify that the "directory" variable should point to the full or relative path of the tide_models directory. While the current example is already good, in my experience, it's always best to assume that some users may not be entirely sure what to do. Providing a more explicit example, such as a full path, could make it clearer.

Just working through the examples now; I'm choosing to test with FES2022 because that's what I've got to hand, but it wasn't immediately clear what level of the directory needed to be provided. In other words, I provided `tide_model = 'FES2022'` but the of course got an error related to there being no model under that name, because `tag_tides()` was looking specifically for `'FES2022_load'` or `'FES2022_extrapolated'`. You've done an amazing job clarifying the often clunky and confusing world of tide constituent sources in the docs; perhaps a very minor addition to explain that the exact string returned by `list_models()` is what needs to be set as `tide_model` would be helpful!
