### 2025-04-04 #7786 @fmemuir
@robbibt FYI I received an error when installing via `pip`, it doesn't seem to have caused any issues so far and might only be relevant for specific dependency versions I already had installed, but wanted to flag anyway:
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed.
This behaviour is the source of the following dependency conflicts.
    gensim 4.3.0 requires FuzzyTM>=0.4.0, which is not installed.
```
