### 2025-04-04 #7786 @daniloceano
Hi @robbibt! I noticed a small inconsistency in the documentation related to the color interpretation of tide heights:

![Image](https://github.com/user-attachments/assets/ef11ec62-1b09-4266-8744-c8fbe8317b42)

The text says:
*"Blue values below indicate low tide pixels, while red indicates high tide pixels..."*

But in the figure, it seems to be the opposite — blue corresponds to **higher** values (around 1.0), and red to **lower** values (around -1.0), based on the colorbar. Maybe the colormap being used is `"RdBu"` instead of `"RdBu_r"` (or vice versa)?
