### 2025-04-04 #7786 @daniloceano
Just a small suggestion — in the example using `.odc.explore()` to visualize statistics on a map, it might be helpful to include a colorbar or some way of interpreting the values being shown:

![Image](https://github.com/user-attachments/assets/e135ecc9-dc60-4b48-8919-4059aac0aba8)

I know this is a bit of a detail, and it may not be essential to the core functionality, but it could really help users better understand what the map is representing at a glance.

If it's too much work to include, no worries — just thought it could be a nice enhancement.
