### 2025-04-11 #7786 @fmemuir
Hi @robbibt, just ran into a few raster analysis errors after leaving the example notebook open on my server machine. I thought it might be related to a week-long timeout contingency within the STAC catalogue service, and restarting the kernel solves the "AppDefined in HTTP response code" error, but the "Ignoring read failure while reading" error remains. However I still get successful side-by-side NDWI composite plots in the cells below. Any chance a 'plain English' wrapper for this error (or just a note in the documentation) might be incorporated, in case anyone else runs into this?

![Image](https://github.com/user-attachments/assets/b92321ee-ce43-49ea-b7bc-33cd31422511)

![Image](https://github.com/user-attachments/assets/55652616-45f1-4f9b-adef-af3cc38bb0e6)
