+++ paper/paper.bib
@@ -77,7 +77,8 @@
   author={Carrere, Loren and Lyard, Florent and Cancet, Mathilde and Allain, Damien and Dabat, Mei-Ling and Fouchet, Ergane and Sahuc, Etienne and Faugere, Yannice and Dibarboure, Gerald and Picot, Nicolas},
   booktitle={2022 Ocean Surface Topography Science Team Meeting},
   pages={43},
-  year={2022}
+  year={2022},
+  doi = {10.24400/527896/a03-2022.3287}
 }
 
 @article{GESLAv3,
@@ -210,6 +211,18 @@
   publisher={Elsevier}
 }
 
+@article{vos2023benchmark,
+  title = {Benchmarking Satellite-Derived Shoreline Mapping Algorithms},
+  author = {Vos, Kilian and Splinter, Kristen D. and {Palomar-V{\'a}zquez}, J. E. and {Almonacid-Caballer}, Jaime and {Cabezas-Rabad{\'a}n}, C. and Kras, Etienne C. and Luijendijk, Arjen P. and Calkoen, Floris and Almeida, Lu{\'i}s Pedro and Pais, D. and Klein, Antonio H.F. and Mao, Y. and Harris, D. and Castelle, Bruno and Buscombe, Daniel and Vitousek, Sean},
+  year = {2023},
+  journal = {Communications Earth \& Environment},
+  volume = {4},
+  number = {345},
+  publisher = {Springer US},
+  doi = {https://doi.org/10.1038/s43247-023-01001-2}
+}
+
+
 @article{sent2025time,
   title={What time is the tide? The importance of tides for ocean colour applications to estuaries},
   author={Sent, G. and Antunes, C. and Spyrakos, E. and Jackson, T. and Atwood, E. C. and Brito, A. C},
@@ -227,3 +240,17 @@ title = {{SpatioTemporal Asset Catalog (STAC) specification}},
 url = {https://stacspec.org},
 year = {2024}
 }
+
+@article{fitton2021intertidal,
+  title = {Remotely Sensed Mapping of the Intertidal Zone: {{A Sentinel-2}} and {{Google Earth Engine}} Methodology},
+  author = {Fitton, James M. and Rennie, Alistair F. and Hansom, James D. and Muir, Freya M. E.},
+  year = {2021},
+  journal = {Remote Sensing Applications: Society and Environment},
+  volume = {22},
+  number = {100499},
+  pages = {1--12},
+  publisher = {Elsevier},
+  issn = {23529385},
+  doi = {10.1016/j.rsase.2021.100499}
+}
+
