### 2025-04-11 #7786 @fmemuir
@pdebuyl I've submitted a pull request with my minor paper edits and comments; once these are addressed (plus the minor documentation/code edits acknowledged by @robbibt above), I'm happy to conclude the review from my side!
