### 2025-04-16 42f2812
Minor paper updates
