+++ paper/paper.bib
@@ -199,18 +199,6 @@
   url={https://doi.org/10.1016/j.coastaleng.2021.103919}
 }
 
-@article{vos2019coastsat,
-  title={{CoastSat}: {A} {Google Earth Engine}-enabled {Python} toolkit to extract shorelines from publicly available satellite imagery},
-  author={Vos, K. and Splinter, K. D. and Harley, M. D. and Simmons, J. A. and Turner, I. L.},
-  journal={Environmental Modelling \& Software},
-  volume={122},
-  pages={104528},
-  year={2019},
-  doi={10.1016/j.envsoft.2019.104528},
-  url={https://doi.org/10.1016/j.envsoft.2019.104528},
-  publisher={Elsevier}
-}
-
 @article{vos2023benchmark,
   title = {Benchmarking Satellite-Derived Shoreline Mapping Algorithms},
   author = {Vos, Kilian and Splinter, Kristen D. and {Palomar-V{\'a}zquez}, J. E. and {Almonacid-Caballer}, Jaime and {Cabezas-Rabad{\'a}n}, C. and Kras, Etienne C. and Luijendijk, Arjen P. and Calkoen, Floris and Almeida, Lu{\'i}s Pedro and Pais, D. and Klein, Antonio H.F. and Mao, Y. and Harris, D. and Castelle, Bruno and Buscombe, Daniel and Vitousek, Sean},
@@ -221,7 +209,6 @@
   publisher = {Springer US},
   doi = {https://doi.org/10.1038/s43247-023-01001-2}
 }
-
 
 @article{sent2025time,
   title={What time is the tide? The importance of tides for ocean colour applications to estuaries},
