### 2025-04-16 969e1d3
JOSS review updates: Minor paper changes (#82)
* Minor paper updates

* Update generated PDF

* Update paper.md

* Add files via upload

* Update generated PDF

* Update paper.bib

* Update generated PDF

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
