+++ paper/paper.bib
@@ -240,4 +240,3 @@ year = {2024}
   issn = {23529385},
   doi = {10.1016/j.rsase.2021.100499}
 }
-
