### 2025-04-22 8cb9137
Update tests to use MPC
