### 2025-04-22 fa0ba44
Use DEA loading as a fallback
