### 2025-04-22 207e44d
Reconfigure setup docs
