### 2025-04-22 b1d751c
JOSS review: Documentation and notebook updates (#83)
* JOSS review: Documentation and notebook updates

* Remove geotiff

* Update git ignore

* Update notebook

* Fix file

* Update generated PDF

* Update tests to use MPC

* Update generated PDF

* Use DEA loading as a fallback

* Update generated PDF

* Reconfigure setup docs

* Update generated PDF

* Minor update

* Update generated PDF

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
