+++ docs/setup.md
@@ -272,33 +272,124 @@ Follow the guides below for some of the most commonly used global ocean tide mod
 
     To allow you to improve tide modelling performance by clipping your tide model files ([see below](#clipping-model-files-to-improve-performance)), we recommend downloading NetCDF-format versions of tide models wherever possible.
 
-## Configuring `eo-tides` to use tide model directory
+## Verifying available and supported models
 
-`eo-tides` can be pointed to the location of your [tide model directory](#setting-up-a-tide-model-directory) and your downloaded tide model data in two ways:
+You can check what tide models have been correctly set up for use by `eo-tides` using [`eo_tides.utils.list_models`](api.md#eo_tides.utils.list_models).
 
-### 1. Using the `directory` function parameter
-
-All tide modelling functions from `eo-tides` provide a `directory` parameter that can be used to specify **either the relative or full/absolute path** to your tide model directory.
-
-For example, using the [`eo_tides.model.model_tides`](api.md#eo_tides.model.model_tides) function:
+The `list_models` function (like all main functions from `eo-tides`) requires you to specify **either the relative or full/absolute path** to your tide model directory using the `directory` parameter:
 
 === "Relative path"
 
-    ```py hl_lines="8"
+    ```py hl_lines="4"
+    from eo_tides.utils import list_models
+
+    available_models, supported_models = list_models(
+        directory="tide_models/"
+    )
+    ```
+
+    Running `list_models` will print out a useful summary table, with supported tide models listed under **"Model"**, and the expected path to their data within your tide model directory listed under **"Expected path"**.
+    Models that are correctly set up and available are marked with a ✅, while unavailable (or incorrectly configured) models are marked with a ❌:
+
+    ```
+    ──────────────────────────────────────────────────────────
+    󠀠🌊  | Model        | Expected path
+    ──────────────────────────────────────────────────────────
+    ✅  │ EOT20        │ tide_models/EOT20/ocean_tides
+    ❌  │ FES2014      │ tide_models/fes2014/ocean_tide
+    ✅  │ HAMTIDE11    │ tide_models/hamtide
+    ❌  │ TPXO9.1      │ tide_models/TPXO9.1/DATA
+    ...   ...            ...
+    ──────────────────────────────────────────────────────────
+
+    Summary:
+    Available models: 2/50
+    ```
+
+=== "Full path (Windows)"
+
+    ```py hl_lines="4"
+    from eo_tides.utils import list_models
+
+    available_models, supported_models = list_models(
+        directory="D:/projects/tide_models/"
+    )
+    ```
+
+    Running `list_models` will print out a useful summary table, with supported tide models listed under **"Model"**, and the expected path to their data within your tide model directory listed under **"Expected path"**.
+    Models that are correctly set up and available are marked with a ✅, while unavailable (or incorrectly configured) models are marked with a ❌:
+
+    ```
+    ──────────────────────────────────────────────────────────
+    󠀠🌊  | Model        | Expected path
+    ──────────────────────────────────────────────────────────
+    ✅  │ EOT20        │ D:/projects/tide_models/EOT20/ocean_tides
+    ❌  │ FES2014      │ D:/projects/tide_models/fes2014/ocean_tide
+    ✅  │ HAMTIDE11    │ D:/projects/tide_models/hamtide
+    ❌  │ TPXO9.1      │ D:/projects/tide_models/TPXO9.1/DATA
+    ...   ...            ...
+    ──────────────────────────────────────────────────────────
+
+    Summary:
+    Available models: 2/50
+    ```
+
+=== "Full path (Linux)"
+
+    ```py hl_lines="4"
[…]
