### 2025-04-23 052d69f
Update setup.md
