### 2025-04-23 #7786 @robbibt
Hi @pdebuyl, I've just finished making some major updates based on the reviews above.

* @daniloceano: Updated all satellite EO notebooks to use globally available Microsoft Planetary Computer STAC API, instead of the Australia-specific DEA STAC API. This should allow the examples to be more easily adapted to global applications.
* @fmemuir: Reconfigured "setup" docs, reducing focus on setting tide directory and explaining how tide model names can be obtained from the `list_models` output (https://geoscienceaustralia.github.io/eo-tides/setup/#verifying-available-and-supported-models)
* @daniloceano: Replaced interactive `.explore()` plots with static matplotlib plots to provide legends and reduce compatibility issues
* @fmemuir: Added link to "contributing.md" to the `eo-tides` readme and docs homepage (https://github.com/GeoscienceAustralia/eo-tides?tab=readme-ov-file#contributing)
* @daniloceano: Fixed inconsistency in the documentation related to the color interpretation of tide heights
* @fmemuir: Merged and incorportated your documentation edits and suggestions (https://github.com/GeoscienceAustralia/eo-tides/pull/81, https://github.com/GeoscienceAustralia/eo-tides/pull/82)

For this one:

> @fmemuir Any chance a 'plain English' wrapper for this error (or just a note in the documentation) might be incorporated, in case anyone else runs into this?

Unfortunately this comes directly from `odc-stac`, and it seems to occur randomly as I think it's an ephemeral cloud data access issue. I'll try and find a way to reproduce this more reliably, and if so raise an issue on the `odc-stac` to see if more graceful error messages could be add there instead.

I think this should hopefully address almost all of the feedback I've recieved so far - thanks so much for all the great suggestions! You can see a full list of JOSS-related PRs and updates made to the repo here:

- https://github.com/GeoscienceAustralia/eo-tides/pulls?q=JOSS+review
- https://github.com/GeoscienceAustralia/eo-tides/issues?q=JOSS+review
