### 2025-04-23 GeoscienceAustralia/eo-tides#79 @robbibt
Done in #83
