### 2025-04-24 #7786 @daniloceano
Hi @pdebuyl ! I was just waiting for the latest updates to check if the program would work well in other regions. I’ve tested the documentation for several regions around the globe, and everything is working great! 😊

I’m happy to have finished my part of the review. Congratulations to @robbibt  and the team for the excellent work and thanks for addressing my issues.
