### 2025-04-25 39f3568
Create dependabot.yml
