+++ paper/paper.bib
@@ -207,7 +207,7 @@
   volume = {4},
   number = {345},
   publisher = {Springer US},
-  doi = {https://doi.org/10.1038/s43247-023-01001-2}
+  doi = {10.1038/s43247-023-01001-2}
 }
 
 @article{sent2025time,
