### 2025-04-28 #7786 @robbibt
> @robbibt I'll proceed to the subsequent editorial steps (proofreading one more, etc) and will keep you posted.

Thanks @pdebuyl, I just merged your DOI PR
