### 2025-05-09 d217f99
Add first draft of NDWI corr code
