+++ eo_tides/validation.py
@@ -1,4 +1,5 @@
 import datetime
+import os
 import warnings
 from math import sqrt
 from numbers import Number
@@ -6,11 +7,15 @@ from numbers import Number
 import geopandas as gpd
 import pandas as pd
 import tqdm
-from odc.geo.geom import BoundingBox
+import xarray as xr
+from odc.geo.geom import BoundingBox, point
 from pandas.tseries.offsets import MonthBegin, MonthEnd, YearBegin, YearEnd
 from scipy import stats
 from shapely.geometry import Point
 from sklearn.metrics import mean_absolute_error, mean_squared_error
+
+from .eo import tag_tides
+from .io import load_ndwi_mpc
 
 
 def eval_metrics(x, y, round=3, all_regress=False):
@@ -332,3 +337,148 @@ def load_gauge_gesla(
 
     # Return data
     return data_df
+
+
+def model_rankings_ndwi(
+    x: float,
+    y: float,
+    time_range: tuple[str, str] = ("2022", "2024"),
+    model: str | list[str] = "all",
+    directory: str | os.PathLike | None = None,
+    crs: str = "EPSG:4326",
+    buffer_radius: float = 2500,
+    ndwi_threshold: float = 0.0,
+    freq_min: float = 0.01,
+    freq_max: float = 0.99,
+    corr_min: float = 0.15,
+    load_ls: bool = True,
+    load_s2: bool = True,
+):
+    """
+    Ranks tide models at a coastal location based on their
+    correlation with satellite-observed inundation patterns
+    derived from the Normalised Difference Water Index (NDWI).
+
+    Correlations are calculated across a buffered region around
+    an input point. High correlations indicate that a tide model
+    correctly sorted satellite imagery by tide height, with
+    high tide observations being consistently wet, and low tide
+    observations being consistently dry.
+
+    Parameters
+    ----------
+    x : float
+        X coordinate of the coastal point of interest. By default
+        this should be in "EPSG:4326" degrees longitude; use "crs"
+        for custom coordinate reference systems.
+    y : float
+        Y coordinate of the coastal point of interest. By default
+        this should be in "EPSG:4326" degrees latitude; use
+        "crs" for custom coordinate reference systems.
+    time_range : tuple, optional
+        The time range to load data for as a tuple of strings (e.g.
+        `("2020", "2021")`. If not provided, data will be loaded for
+        all available timesteps.
+    model : str or list of str, optional
+        The tide model (or list of models) to compare. Defaults to
+        "all", which will compare all models available in `directory`.
+        For a full list of available and supported models,
+        run `from eo_tides.utils import list_models; list_models()`.
+    directory : str, optional
+        The directory containing tide model data files. If no path is
+        provided, this will default to the environment variable
+        `EO_TIDES_TIDE_MODELS` if set, or raise an error if not.
+        Tide modelling files should be stored in sub-folders for each
+        model that match the structure required by `pyTMD`
+        (<https://geoscienceaustralia.github.io/eo-tides/setup/>).
+    crs : str, optional
+        Input coordinate reference system for x and y coordinates.
+        Defaults to "EPSG:4326" (WGS84; degrees latitude, longitude).
+    buffer_radius : float, optional
+        Radius in meters for generating circular buffer around the
+        input point.
+    ndwi_threshold : float, optional
+        NDWI threshold above which a pixel is considered wet.
+    freq_min : float, optional
+        Minimum fraction of time a pixel must be wet to be considered
+        intertidal and included in the analysis.
+    freq_max : float, optional
+        Maximum fraction of time a pixel can be wet to be considered
[…]
