### 2025-05-12 #7786 @pdebuyl
Hello @robbibt I am actually always installing and testing the code in the articles I edit. Installation of eo-tides and of a sample model worked flawlessly. The documentation is very nice.

What remains to be done:
1. If Vanessa Newey has an ORCID, it would be good to have it. It is not mandatory though.
2. archive a release of eo-tides on zenodo and share the specific version and DOI here
3. I have no additional comment on the manuscript itself.
