### 2025-05-12 #7786 @robbibt
Thanks so much @pdebuyl - appreciate you testing it out as well!

Quick question:
Geoscience Australia (the Australian government agency where I work) has [our own repository](https://ecat.ga.gov.au/geonetwork/srv/eng/catalog.search#/search?query_string=%7B%22resourceType%22:%20%7B%22software%22:%20true%7D%7D) where we can archive code and mint a DOI. Would it be possible for us to use that instead of Zenodo? No worries either way.
