### 2025-05-13 #7786 @pdebuyl
Thanks for adding the ORCID! Regarding the repository, it is possible to use something else than zenodo. As it is the first time I have this case, I will look into whether we have specific criteria for the type of archive. I'll update you here shortly.
