### 2025-05-13 #7786 @robbibt
There's some more info about the portal here if it helps: https://www.ga.gov.au/data-pubs/datastandards/cataloguestandard
https://www.ga.gov.au/data-pubs/datastandards/fairdata
