### 2025-05-13 #7786 @pdebuyl
hello @robbibt I found no restriction in our docs and also previous JOSS papers where the archive was in an institutional repository, please proceed with the solution you proposed
