### 2025-05-14 4a29ce0
Version 0.7.0 release for JOSS archiving
