### 2025-05-14 7bfc54f
Version 0.7.0 release for JOSS archiving (#91)
