### 2025-05-14 #7786 @robbibt
Hi @pdebuyl, here's the persistent DOI for our archived code!
https://doi.org/10.26186/150065

If you scroll down, you can see the archived `.zip` containing all code from `eo-tides==0.7.0` (the latest version as of this morning).

![Image](https://github.com/user-attachments/assets/4f1fd20b-7ef7-4cf7-bce5-fd19b8060e7c)
