+++ .github/actions/setup-python-env/action.yml
@@ -13,7 +13,7 @@ inputs:
   uv-version:
     description: "uv version to use"
     required: true
-    default: "0.6.2"
+    default: "0.7.0"
 
 runs:
   using: "composite"
