+++ .github/workflows/publish.yml
@@ -7,7 +7,6 @@ on:
     branches: [main]
 
 jobs:
-
   # Create a relevant package version for releases and dev changes
   set-version:
     runs-on: ubuntu-latest
