+++ tests/test_validation.py
@@ -45,7 +45,7 @@ def test_load_gauge_gesla(x, y, site_code, max_distance, correct_mean, expected)
         max_distance=max_distance,
         correct_mean=correct_mean,
         time=("2018-01-01", "2018-01-20"),
-        data_path="tests/data/",
+        data_path="tests/data/GESLA3.0_ALL/",
         metadata_path="tests/data/GESLA3_ALL 2.csv",
     )
 
