### 2025-05-20 4921768
Revert pre-commit
