### 2025-05-22 #7786 @robbibt
Hi @pdebuyl, I was just wondering if there was an update on the final publication - no worries if not, but we have an Australia coastal conference early next month where I'll be giving a talk about `eo-tides`, and I'd really love to point users to the published JOSS paper if at all possible!
