### 2025-05-22 fcaf2a4
Restore precommit, pin to 0.7.5
