+++ .github/actions/setup-python-env/action.yml
@@ -13,7 +13,7 @@ inputs:
   uv-version:
     description: "uv version to use"
     required: true
-    default: "0.7.0"
+    default: "0.7.5"
 
 runs:
   using: "composite"
