+++ .pre-commit-config.yaml
@@ -26,7 +26,6 @@ repos:
     hooks:
       - id: prettier
 
-repos:
   - repo: https://github.com/astral-sh/uv-pre-commit
     # uv version.
     rev: 0.7.5
