### 2025-05-22 f1ce3a4
Simplify publish action to use dynamic versioning
