+++ .github/workflows/publish.yml
@@ -1,4 +1,4 @@
-name: Build docs and publish main
+name: Build and publish package and docs
 
 on:
   release:
@@ -18,6 +18,7 @@ jobs:
         with:
           fetch-depth: 0
 
+      # Set up environment using `uv`
       - name: Set up the environment
         uses: ./.github/actions/setup-python-env
 
