### 2025-05-22 b8cb8fc
Add new update changelog workflow
