### 2025-05-22 90bcacd
Use dynamic version handling via `hatch-vcs`, add automatic changelog update action (#95)
* Test vcs version, precommit for lockfile

* Revert pre-commit

* Restore precommit, pin to 0.7.5

* Fix precommit format

* Update lock file

* Simplify publish action to use dynamic versioning

* Update doco

* Add new update changelog workflow

* Update changelog format

* Update changelog
