+++ .github/workflows/publish.yml
@@ -1,4 +1,4 @@
-name: Build docs and publish main
+name: Build and publish package and docs
 
 on:
   release:
@@ -7,64 +7,27 @@ on:
     branches: [main]
 
 jobs:
-  # Create a relevant package version for releases and dev changes
-  set-version:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0 # Need full history for tags
-
-      - name: Export tag
-        if: ${{ github.event_name == 'release' }}
-        id: vars
-        run: echo tag=${GITHUB_REF#refs/*/} >> $GITHUB_OUTPUT
-
-      # For full releases, use tag for version name
-      - name: Set release version
-        if: ${{ github.event_name == 'release' }}
-        run: |
-          sed -i "s/^version = \".*\"/version = \"$RELEASE_VERSION\"/" pyproject.toml
-        env:
-          RELEASE_VERSION: ${{ steps.vars.outputs.tag }}
-
-      # For merges to main, incremement patch version by one
-      # and use ".devN" suffix based on commits since tag
-      - name: Set dev version
-        if: ${{ github.event_name == 'push' }}
-        run: |
-          LATEST_TAG=$(git describe --tags --abbrev=0)
-          IFS='.' read -r MAJOR MINOR PATCH <<< "$LATEST_TAG"
-          PATCH=$((PATCH + 1))
-          DEV_NUM=$(git rev-list $LATEST_TAG..HEAD --count)
-          DEV_VERSION="$MAJOR.$MINOR.$PATCH.dev$DEV_NUM"
-          sed -i "s/^version = \".*\"/version = \"$DEV_VERSION\"/" pyproject.toml
-
-      - name: Upload updated pyproject.toml
-        uses: actions/upload-artifact@v4
-        with:
-          name: pyproject-toml
-          path: pyproject.toml
-
   # Publish package to PyPI
   publish:
     runs-on: ubuntu-latest
-    needs: [set-version]
     steps:
+      # Check out Git repository with full history
+      # Required for `hatchling-vcs` to get correct version
       - name: Check out
         uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
 
+      # Set up environment using `uv`
       - name: Set up the environment
         uses: ./.github/actions/setup-python-env
 
-      - name: Download updated pyproject.toml
-        uses: actions/download-artifact@v4
-        with:
-          name: pyproject-toml
-
+      # Package will be built using version automatically
+      # generated from Git by `hatchling-vcs`
       - name: Build package
         run: uv build
 
+      # Publish to PyPI
       - name: Publish package
         run: uv publish
         env:
