+++ pyproject.toml
@@ -94,6 +94,9 @@ build-backend = "hatchling.build"
 [tool.hatch.version]
 source = "vcs"
 
+[tool.hatch.version.raw-options]
+local_scheme = "no-local-version"
+
 [tool.mypy]
 files = ["eo_tides"]
 python_version = "3.10"
