### 2025-05-25 #7786 @pdebuyl
> [@robbibt](https://github.com/robbibt) I may have missed it but does your software archive specify the version of the software associated with it?

Hello @kthyng indeed I did not find a version on the repository itself, and relied on @robbibt 's input for this information.
