### 2025-05-25 69518bb
Fix capitalisation in JOSS paper bibliography
