+++ paper/paper.bib
@@ -20,7 +20,7 @@
 
 @InProceedings{ mckinney-proc-scipy-2010,
   author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  title     = { {D}ata structures for statistical computing in {P}ython },
   booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
   pages     = { 56 - 61 },
   year      = { 2010 },
@@ -50,7 +50,7 @@
 }
 
 @article{murray2012continental,
-  title={Continental scale mapping of tidal flats across East Asia using the Landsat archive},
+  title={Continental scale mapping of tidal flats across {East Asia} using the {Landsat} archive},
   author={Murray, Nicholas J and Phinn, Stuart R and Clemens, Robert S and Roelfsema, Chris M and Fuller, Richard A},
   journal={Remote Sensing},
   volume={4},
@@ -62,7 +62,7 @@
 }
 
 @article{sagar2017item,
-  title={Extracting the intertidal extent and topography of the Australian coastline from a 28 year time series of Landsat observations},
+  title={Extracting the intertidal extent and topography of the {Australian} coastline from a 28 year time series of {Landsat} observations},
   author={Sagar, S. and Roberts, D. and Bala, B. and Lymburner, L.},
   journal={Remote Sensing of Environment},
   volume={195},
@@ -140,7 +140,7 @@
 }
 
 @article{bishop2021mapping,
-  title={Mapping {Australia}'s {Dynamic Coastline} at {Mean Sea Level} Using {Three Decades of Landsat Imagery}},
+  title={Mapping {Australia}'s dynamic coastline at {Mean Sea Level} using three decades of {Landsat} imagery},
   author={Bishop-Taylor, R. and Nanson, R. and Sagar, S. and Lymburner, L.},
   journal={Remote Sensing of Environment},
   volume={267},
@@ -152,7 +152,7 @@
 }
 
 @article{bishop2019NIDEM,
-  title={Between the tides: Modelling the elevation of Australia's exposed intertidal zone at continental scale},
+  title={Between the tides: Modelling the elevation of {Australia}'s exposed intertidal zone at continental scale},
   author={Bishop-Taylor, R. and Sagar, S. and Lymburner, L. and Beaman, R.J.},
   journal={Estuarine, Coastal and Shelf Science},
   volume={223},
@@ -164,7 +164,7 @@
 }
 
 @article{sagar2018composites,
-  title={Generating Continental Scale Pixel-Based Surface Reflectance Composites in Coastal Regions with the Use of a Multi-Resolution Tidal Model},
+  title={Generating continental scale pixel-based surface reflectance composites in coastal regions with the use of a multi-resolution tidal model},
   author={Sagar, S. and Phillips, C. and Bala, B. and Roberts, D. and Lymburner, L. and Beaman, R.J.},
   journal={Remote Sensing},
   volume={10(3)},
@@ -200,7 +200,7 @@
 }
 
 @article{vos2023benchmark,
-  title = {Benchmarking Satellite-Derived Shoreline Mapping Algorithms},
+  title = {Benchmarking satellite-derived shoreline mapping algorithms},
[…]
