+++ paper/paper.bib
@@ -34,7 +34,7 @@
   journal = {Journal of Open Research Software},
   month = apr,
   number = {1},
-  title = {{xarray}: N-D labeled Arrays and Datasets in Python},
+  title = {{xarray}: {N-D} labeled arrays and datasets in {Python}},
   volume = {5},
   year = {2017}
 }
