### 2025-05-25 06da377
Fix capitalisation in JOSS paper bibliography (#97)
* Fix capitalisation in JOSS paper bibliography

* Update generated PDF

* One more fix

* Update generated PDF

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
