### 2025-05-25 #7786 @robbibt
Hi @kthyng, we've updated our institutional repo to clearly describe the version:

![Image](https://github.com/user-attachments/assets/5d653c95-446f-412b-87e4-3acace5759f0)

I've also updated the capitalisation throughout the bibliography - please let me know if you notice any remaining issues.
