### 2025-05-27 be1dd92
Add version attribute to init file
