+++ .github/release.yml
@@ -11,6 +11,9 @@ changelog:
     - title: Bug fixes
       labels:
         - bug
+    - title: Documentation updates
+      labels:
+        - documentation
     - title: Other changes
       labels:
         - "*"
