+++ .github/dependabot.yml
@@ -13,3 +13,6 @@ updates:
       python-deps:
         patterns:
           - "*"
+    labels:
+      - "dependencies"
+      - "packaging"
