### 2025-05-29 00b4707
Update citation to match JOSS paper
