+++ CITATION.cff
@@ -1,29 +1,5 @@
 cff-version: "1.2.0"
 authors:
-- family-names: Bishop-Taylor
-  given-names: Robbi
-  orcid: "https://orcid.org/0000-0002-1533-2599"
-- family-names: Phillips
-  given-names: Claire
-  orcid: "https://orcid.org/0009-0003-9882-9131"
-- family-names: Sagar
-  given-names: Stephen
-  orcid: "https://orcid.org/0000-0001-9568-9661"
-- family-names: Newey
-  given-names: Vanessa
-  orcid: "https://orcid.org/0000-0003-3705-5665"
-- family-names: Sutterley
-  given-names: Tyler
-  orcid: "https://orcid.org/0000-0002-6964-1194"
-contact:
-- family-names: Bishop-Taylor
-  given-names: Robbi
-  orcid: "https://orcid.org/0000-0002-1533-2599"
-doi: 10.26186/150065
-message: If you use this software, please cite our article in the
-  Journal of Open Source Software.
-preferred-citation:
-  authors:
   - family-names: Bishop-Taylor
     given-names: Robbi
     orcid: "https://orcid.org/0000-0002-1533-2599"
@@ -39,6 +15,30 @@ preferred-citation:
   - family-names: Sutterley
     given-names: Tyler
     orcid: "https://orcid.org/0000-0002-6964-1194"
+contact:
+  - family-names: Bishop-Taylor
+    given-names: Robbi
+    orcid: "https://orcid.org/0000-0002-1533-2599"
+doi: 10.26186/150065
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+    - family-names: Bishop-Taylor
+      given-names: Robbi
+      orcid: "https://orcid.org/0000-0002-1533-2599"
+    - family-names: Phillips
+      given-names: Claire
+      orcid: "https://orcid.org/0009-0003-9882-9131"
+    - family-names: Sagar
+      given-names: Stephen
+      orcid: "https://orcid.org/0000-0001-9568-9661"
+    - family-names: Newey
+      given-names: Vanessa
+      orcid: "https://orcid.org/0000-0003-3705-5665"
+    - family-names: Sutterley
+      given-names: Tyler
+      orcid: "https://orcid.org/0000-0002-6964-1194"
   date-published: 2025-05-29
   doi: 10.21105/joss.07786
   issn: 2475-9066
