### 2025-05-30 6871a80
Update suggested citation to use new JOSS paper citation (#100)
* Update citation to match JOSS paper

* Update citation information

* Fix citation format
