### 2025-05-30 5b46d21
Reformat code with additional `ruff` linting rules (#101)
* Update ruff linting

* Update generated PDF

* Fix set literals

* Update generated PDF

* Update paper generation action

---------

Co-authored-by: robbibt <robbibt@users.noreply.github.com>
