+++ tests/test_eo.py
@@ -358,7 +358,7 @@ def test_pixel_tides_ensemble(satellite_ds):
         ensemble_models=ENSEMBLE_MODELS,
     )
 
-    assert set(modelled_tides_ds.tide_model.values) == set([
+    assert set(modelled_tides_ds.tide_model.values) == {
         "EOT20",
         "HAMTIDE11",
         "ensemble-best",
@@ -366,4 +366,4 @@ def test_pixel_tides_ensemble(satellite_ds):
         "ensemble-mean-top2",
         "ensemble-mean-weighted",
         "ensemble-mean",
-    ])
+    }
