+++ .github/workflows/update-changelog.yml
@@ -32,10 +32,9 @@ jobs:
           title: "Update changelog for ${{ github.event.release.tag_name }}"
           body: |
             This PR updates the changelog with details from the release [${{ github.event.release.tag_name }}](${{ github.event.release.html_url }}).
-          commit-message: "Update changelog for ${{ github.event.release.tag_name }} ["
+          commit-message: "Update changelog for ${{ github.event.release.tag_name }} [skip ci]"
           branch: update-changelog-${{ github.event.release.tag_name }}
           base: ${{ github.event.release.target_commitish }}
           add-paths: |
             docs/changelog.md
-            uv.lock
           delete-branch: true
