### 2025-05-30 b001cdc
Ensure bare URLs in changelog are correctly linked
