+++ mkdocs.yml
@@ -117,6 +117,7 @@ markdown_extensions:
       generic: true
   - admonition
   - pymdownx.details
+  - pymdownx.magiclink
   - pymdownx.tabbed:
       alternate_style: true
   - pymdownx.superfences
