### 2025-05-30 67c476e
Exclude changelog from pre-commit checks
