+++ .pre-commit-config.yaml
@@ -1,3 +1,5 @@
+exclude: ^docs/changelog\.md$
+
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: "v4.4.0"
@@ -8,6 +10,8 @@ repos:
       - id: check-yaml
       - id: end-of-file-fixer
       - id: trailing-whitespace
+    exclude:
+      - "docs/changelog.md"
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.6.4
