+++ .pre-commit-config.yaml
@@ -10,8 +10,6 @@ repos:
       - id: check-yaml
       - id: end-of-file-fixer
       - id: trailing-whitespace
-    exclude:
-      - "docs/changelog.md"
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.6.4
