### 2025-05-30 fac8d3d
Fix unlinked URLs in changelog (#103)
* Ensure bare URLs in changelog are correctly linked

* Exclude changelog from pre-commit checks

* Update .pre-commit-config.yaml
