+++ .pre-commit-config.yaml
@@ -1,3 +1,5 @@
+exclude: ^docs/changelog\.md$
+
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: "v4.4.0"
