### 2025-05-30 b6f0c3e
Add link to JOSS paper
