+++ docs/index.md
@@ -41,7 +41,7 @@ For instructions on how to set up these models for use in `eo-tides`, refer to [
 
 ## Citing `eo-tides`
 
-To cite `eo-tides` in your work, please use the following Journal of Open Source Software citation:
+To cite `eo-tides` in your work, please use the following [Journal of Open Source Software](https://doi.org/10.21105/joss.07786) citation:
 
 === "Plain text"
 
