+++ README.md
@@ -55,7 +55,7 @@ Interactive Jupyter Notebook usage examples and more complex coastal EO case stu
 
 ## Citing `eo-tides`
 
-To cite `eo-tides` in your work, please use the following Journal of Open Source Software citation:
+To cite `eo-tides` in your work, please use the following [Journal of Open Source Software](https://doi.org/10.21105/joss.07786) citation:
 
 ```
 Bishop-Taylor, R., Phillips, C., Sagar, S., Newey, V., & Sutterley, T., (2025). eo-tides: Tide modelling tools for large-scale satellite Earth observation analysis. Journal of Open Source Software, 10(109), 7786, https://doi.org/10.21105/joss.07786
