### 2025-05-30 1f8fee8
Update JOSS link in readme and docs
